Module Cephalopod_models

Models generated from lexicons

module Base = Cephalopod_xrpc.Base

val all_records : (string, Base.any_record_def) Stdlib.Hashtbl.t

module Types : sig ... end

Type definitions

module Com_Atproto_Temp_RequestPhoneVerification : sig ... end

lexicon "com.atproto.temp.requestPhoneVerification"

module Com_Atproto_Temp_FetchLabels : sig ... end

lexicon "com.atproto.temp.fetchLabels"

module Com_Atproto_Temp_CheckSignupQueue : sig ... end

lexicon "com.atproto.temp.checkSignupQueue"

module Com_Atproto_Temp_AddReservedHandle : sig ... end

lexicon "com.atproto.temp.addReservedHandle"

module Com_Atproto_Sync_SubscribeRepos : sig ... end

lexicon "com.atproto.sync.subscribeRepos"

module Com_Atproto_Sync_RequestCrawl : sig ... end

lexicon "com.atproto.sync.requestCrawl"

module Com_Atproto_Sync_NotifyOfUpdate : sig ... end

lexicon "com.atproto.sync.notifyOfUpdate"

module Com_Atproto_Sync_ListReposByCollection : sig ... end

lexicon "com.atproto.sync.listReposByCollection"

module Com_Atproto_Sync_ListRepos : sig ... end

lexicon "com.atproto.sync.listRepos"

module Com_Atproto_Sync_ListHosts : sig ... end

lexicon "com.atproto.sync.listHosts"

module Com_Atproto_Sync_ListBlobs : sig ... end

lexicon "com.atproto.sync.listBlobs"

module Com_Atproto_Sync_GetRepoStatus : sig ... end

lexicon "com.atproto.sync.getRepoStatus"

module Com_Atproto_Sync_GetRepo : sig ... end

lexicon "com.atproto.sync.getRepo"

module Com_Atproto_Sync_GetRecord : sig ... end

lexicon "com.atproto.sync.getRecord"

module Com_Atproto_Sync_GetLatestCommit : sig ... end

lexicon "com.atproto.sync.getLatestCommit"

module Com_Atproto_Sync_GetHostStatus : sig ... end

lexicon "com.atproto.sync.getHostStatus"

module Com_Atproto_Sync_GetHead : sig ... end

lexicon "com.atproto.sync.getHead"

module Com_Atproto_Sync_GetCheckout : sig ... end

lexicon "com.atproto.sync.getCheckout"

module Com_Atproto_Sync_GetBlocks : sig ... end

lexicon "com.atproto.sync.getBlocks"

module Com_Atproto_Sync_GetBlob : sig ... end

lexicon "com.atproto.sync.getBlob"

module Com_Atproto_Sync_Defs : sig ... end

lexicon "com.atproto.sync.defs"

module Com_Atproto_Server_UpdateEmail : sig ... end

lexicon "com.atproto.server.updateEmail"

module Com_Atproto_Server_RevokeAppPassword : sig ... end

lexicon "com.atproto.server.revokeAppPassword"

module Com_Atproto_Server_ResetPassword : sig ... end

lexicon "com.atproto.server.resetPassword"

module Com_Atproto_Server_ReserveSigningKey : sig ... end

lexicon "com.atproto.server.reserveSigningKey"

module Com_Atproto_Server_RequestPasswordReset : sig ... end

lexicon "com.atproto.server.requestPasswordReset"

module Com_Atproto_Server_RequestEmailUpdate : sig ... end

lexicon "com.atproto.server.requestEmailUpdate"

module Com_Atproto_Server_RequestEmailConfirmation : sig ... end

lexicon "com.atproto.server.requestEmailConfirmation"

module Com_Atproto_Server_RequestAccountDelete : sig ... end

lexicon "com.atproto.server.requestAccountDelete"

module Com_Atproto_Server_RefreshSession : sig ... end

lexicon "com.atproto.server.refreshSession"

module Com_Atproto_Server_ListAppPasswords : sig ... end

lexicon "com.atproto.server.listAppPasswords"

module Com_Atproto_Server_GetSession : sig ... end

lexicon "com.atproto.server.getSession"

module Com_Atproto_Server_GetServiceAuth : sig ... end

lexicon "com.atproto.server.getServiceAuth"

module Com_Atproto_Server_GetAccountInviteCodes : sig ... end

lexicon "com.atproto.server.getAccountInviteCodes"

module Com_Atproto_Server_DescribeServer : sig ... end

lexicon "com.atproto.server.describeServer"

module Com_Atproto_Server_DeleteSession : sig ... end

lexicon "com.atproto.server.deleteSession"

module Com_Atproto_Server_DeleteAccount : sig ... end

lexicon "com.atproto.server.deleteAccount"

module Com_Atproto_Server_DeactivateAccount : sig ... end

lexicon "com.atproto.server.deactivateAccount"

module Com_Atproto_Server_CreateSession : sig ... end

lexicon "com.atproto.server.createSession"

module Com_Atproto_Server_CreateInviteCodes : sig ... end

lexicon "com.atproto.server.createInviteCodes"

module Com_Atproto_Server_CreateInviteCode : sig ... end

lexicon "com.atproto.server.createInviteCode"

module Com_Atproto_Server_CreateAppPassword : sig ... end

lexicon "com.atproto.server.createAppPassword"

module Com_Atproto_Server_CreateAccount : sig ... end

lexicon "com.atproto.server.createAccount"

module Com_Atproto_Server_ConfirmEmail : sig ... end

lexicon "com.atproto.server.confirmEmail"

module Com_Atproto_Server_CheckAccountStatus : sig ... end

lexicon "com.atproto.server.checkAccountStatus"

module Com_Atproto_Server_ActivateAccount : sig ... end

lexicon "com.atproto.server.activateAccount"

module Com_Atproto_Repo_UploadBlob : sig ... end

lexicon "com.atproto.repo.uploadBlob"

module Com_Atproto_Repo_PutRecord : sig ... end

lexicon "com.atproto.repo.putRecord"

module Com_Atproto_Repo_ListRecords : sig ... end

lexicon "com.atproto.repo.listRecords"

module Com_Atproto_Repo_ListMissingBlobs : sig ... end

lexicon "com.atproto.repo.listMissingBlobs"

module Com_Atproto_Repo_ImportRepo : sig ... end

lexicon "com.atproto.repo.importRepo"

module Com_Atproto_Repo_GetRecord : sig ... end

lexicon "com.atproto.repo.getRecord"

module Com_Atproto_Repo_DescribeRepo : sig ... end

lexicon "com.atproto.repo.describeRepo"

module Com_Atproto_Repo_DeleteRecord : sig ... end

lexicon "com.atproto.repo.deleteRecord"

module Com_Atproto_Repo_CreateRecord : sig ... end

lexicon "com.atproto.repo.createRecord"

module Com_Atproto_Repo_ApplyWrites : sig ... end

lexicon "com.atproto.repo.applyWrites"

module Com_Atproto_Repo_Defs : sig ... end

lexicon "com.atproto.repo.defs"

module Com_Atproto_Moderation_CreateReport : sig ... end

lexicon "com.atproto.moderation.createReport"

module Com_Atproto_Lexicon_Schema : sig ... end

lexicon "com.atproto.lexicon.schema"

module Com_Atproto_Label_SubscribeLabels : sig ... end

lexicon "com.atproto.label.subscribeLabels"

module Com_Atproto_Label_QueryLabels : sig ... end

lexicon "com.atproto.label.queryLabels"

module Com_Atproto_Identity_UpdateHandle : sig ... end

lexicon "com.atproto.identity.updateHandle"

module Com_Atproto_Identity_SubmitPlcOperation : sig ... end

lexicon "com.atproto.identity.submitPlcOperation"

module Com_Atproto_Identity_SignPlcOperation : sig ... end

lexicon "com.atproto.identity.signPlcOperation"

module Com_Atproto_Identity_ResolveIdentity : sig ... end

lexicon "com.atproto.identity.resolveIdentity"

module Com_Atproto_Identity_ResolveHandle : sig ... end

lexicon "com.atproto.identity.resolveHandle"

module Com_Atproto_Identity_ResolveDid : sig ... end

lexicon "com.atproto.identity.resolveDid"

module Com_Atproto_Identity_RequestPlcOperationSignature : sig ... end

lexicon "com.atproto.identity.requestPlcOperationSignature"

module Com_Atproto_Identity_RefreshIdentity : sig ... end

lexicon "com.atproto.identity.refreshIdentity"

module Com_Atproto_Identity_GetRecommendedDidCredentials : sig ... end

lexicon "com.atproto.identity.getRecommendedDidCredentials"

module Com_Atproto_Identity_Defs : sig ... end

lexicon "com.atproto.identity.defs"

module Com_Atproto_Admin_UpdateSubjectStatus : sig ... end

lexicon "com.atproto.admin.updateSubjectStatus"

module Com_Atproto_Admin_UpdateAccountSigningKey : sig ... end

lexicon "com.atproto.admin.updateAccountSigningKey"

module Com_Atproto_Admin_UpdateAccountPassword : sig ... end

lexicon "com.atproto.admin.updateAccountPassword"

module Com_Atproto_Admin_UpdateAccountHandle : sig ... end

lexicon "com.atproto.admin.updateAccountHandle"

module Com_Atproto_Admin_UpdateAccountEmail : sig ... end

lexicon "com.atproto.admin.updateAccountEmail"

module Com_Atproto_Admin_SendEmail : sig ... end

lexicon "com.atproto.admin.sendEmail"

module Com_Atproto_Admin_SearchAccounts : sig ... end

lexicon "com.atproto.admin.searchAccounts"

module Com_Atproto_Admin_GetSubjectStatus : sig ... end

lexicon "com.atproto.admin.getSubjectStatus"

module Com_Atproto_Admin_GetInviteCodes : sig ... end

lexicon "com.atproto.admin.getInviteCodes"

module Com_Atproto_Admin_GetAccountInfos : sig ... end

lexicon "com.atproto.admin.getAccountInfos"

module Com_Atproto_Admin_GetAccountInfo : sig ... end

lexicon "com.atproto.admin.getAccountInfo"

module Com_Atproto_Admin_EnableAccountInvites : sig ... end

lexicon "com.atproto.admin.enableAccountInvites"

module Com_Atproto_Admin_DisableInviteCodes : sig ... end

lexicon "com.atproto.admin.disableInviteCodes"

module Com_Atproto_Admin_DisableAccountInvites : sig ... end

lexicon "com.atproto.admin.disableAccountInvites"

module Com_Atproto_Admin_DeleteAccount : sig ... end

lexicon "com.atproto.admin.deleteAccount"

module Com_Atproto_Admin_Defs : sig ... end

lexicon "com.atproto.admin.defs"

module Com_Atproto_Server_Defs : sig ... end

lexicon "com.atproto.server.defs"

module Chat_Bsky_Moderation_UpdateActorAccess : sig ... end

lexicon "chat.bsky.moderation.updateActorAccess"

module Chat_Bsky_Moderation_GetMessageContext : sig ... end

lexicon "chat.bsky.moderation.getMessageContext"

module Chat_Bsky_Moderation_GetActorMetadata : sig ... end

lexicon "chat.bsky.moderation.getActorMetadata"

module Chat_Bsky_Convo_UpdateRead : sig ... end

lexicon "chat.bsky.convo.updateRead"

module Chat_Bsky_Convo_UpdateAllRead : sig ... end

lexicon "chat.bsky.convo.updateAllRead"

module Chat_Bsky_Convo_UnmuteConvo : sig ... end

lexicon "chat.bsky.convo.unmuteConvo"

module Chat_Bsky_Convo_SendMessageBatch : sig ... end

lexicon "chat.bsky.convo.sendMessageBatch"

module Chat_Bsky_Convo_SendMessage : sig ... end

lexicon "chat.bsky.convo.sendMessage"

module Chat_Bsky_Convo_RemoveReaction : sig ... end

lexicon "chat.bsky.convo.removeReaction"

module Chat_Bsky_Convo_MuteConvo : sig ... end

lexicon "chat.bsky.convo.muteConvo"

module Chat_Bsky_Convo_ListConvos : sig ... end

lexicon "chat.bsky.convo.listConvos"

module Chat_Bsky_Convo_LeaveConvo : sig ... end

lexicon "chat.bsky.convo.leaveConvo"

module Chat_Bsky_Convo_GetMessages : sig ... end

lexicon "chat.bsky.convo.getMessages"

module Chat_Bsky_Convo_GetLog : sig ... end

lexicon "chat.bsky.convo.getLog"

module Chat_Bsky_Convo_GetConvoForMembers : sig ... end

lexicon "chat.bsky.convo.getConvoForMembers"

module Chat_Bsky_Convo_GetConvoAvailability : sig ... end

lexicon "chat.bsky.convo.getConvoAvailability"

module Chat_Bsky_Convo_GetConvo : sig ... end

lexicon "chat.bsky.convo.getConvo"

module Chat_Bsky_Convo_DeleteMessageForSelf : sig ... end

lexicon "chat.bsky.convo.deleteMessageForSelf"

module Chat_Bsky_Convo_AddReaction : sig ... end

lexicon "chat.bsky.convo.addReaction"

module Chat_Bsky_Convo_Defs : sig ... end

lexicon "chat.bsky.convo.defs"

module Chat_Bsky_Convo_AcceptConvo : sig ... end

lexicon "chat.bsky.convo.acceptConvo"

module Chat_Bsky_Actor_ExportAccountData : sig ... end

lexicon "chat.bsky.actor.exportAccountData"

module Chat_Bsky_Actor_DeleteAccount : sig ... end

lexicon "chat.bsky.actor.deleteAccount"

module Chat_Bsky_Actor_Defs : sig ... end

lexicon "chat.bsky.actor.defs"

module Chat_Bsky_Actor_Declaration : sig ... end

lexicon "chat.bsky.actor.declaration"

module App_Bsky_Video_UploadVideo : sig ... end

lexicon "app.bsky.video.uploadVideo"

module App_Bsky_Video_GetUploadLimits : sig ... end

lexicon "app.bsky.video.getUploadLimits"

module App_Bsky_Video_GetJobStatus : sig ... end

lexicon "app.bsky.video.getJobStatus"

module App_Bsky_Video_Defs : sig ... end

lexicon "app.bsky.video.defs"

module App_Bsky_Unspecced_SearchStarterPacksSkeleton : sig ... end

lexicon "app.bsky.unspecced.searchStarterPacksSkeleton"

module App_Bsky_Unspecced_SearchPostsSkeleton : sig ... end

lexicon "app.bsky.unspecced.searchPostsSkeleton"

module App_Bsky_Unspecced_SearchActorsSkeleton : sig ... end

lexicon "app.bsky.unspecced.searchActorsSkeleton"

module App_Bsky_Unspecced_GetTrendsSkeleton : sig ... end

lexicon "app.bsky.unspecced.getTrendsSkeleton"

module App_Bsky_Unspecced_GetTrends : sig ... end

lexicon "app.bsky.unspecced.getTrends"

module App_Bsky_Unspecced_GetTrendingTopics : sig ... end

lexicon "app.bsky.unspecced.getTrendingTopics"

module App_Bsky_Unspecced_GetTaggedSuggestions : sig ... end

lexicon "app.bsky.unspecced.getTaggedSuggestions"

module App_Bsky_Unspecced_GetSuggestionsSkeleton : sig ... end

lexicon "app.bsky.unspecced.getSuggestionsSkeleton"

module App_Bsky_Unspecced_GetSuggestedUsersSkeleton : sig ... end

lexicon "app.bsky.unspecced.getSuggestedUsersSkeleton"

module App_Bsky_Unspecced_GetSuggestedUsers : sig ... end

lexicon "app.bsky.unspecced.getSuggestedUsers"

module App_Bsky_Unspecced_GetSuggestedStarterPacksSkeleton : sig ... end

lexicon "app.bsky.unspecced.getSuggestedStarterPacksSkeleton"

module App_Bsky_Unspecced_GetSuggestedStarterPacks : sig ... end

lexicon "app.bsky.unspecced.getSuggestedStarterPacks"

module App_Bsky_Unspecced_GetSuggestedFeedsSkeleton : sig ... end

lexicon "app.bsky.unspecced.getSuggestedFeedsSkeleton"

module App_Bsky_Unspecced_GetSuggestedFeeds : sig ... end

lexicon "app.bsky.unspecced.getSuggestedFeeds"

module App_Bsky_Unspecced_GetPostThreadV2 : sig ... end

lexicon "app.bsky.unspecced.getPostThreadV2"

module App_Bsky_Unspecced_GetPostThreadHiddenV2 : sig ... end

lexicon "app.bsky.unspecced.getPostThreadHiddenV2"

module App_Bsky_Unspecced_GetPopularFeedGenerators : sig ... end

lexicon "app.bsky.unspecced.getPopularFeedGenerators"

module App_Bsky_Unspecced_GetConfig : sig ... end

lexicon "app.bsky.unspecced.getConfig"

module App_Bsky_Unspecced_Defs : sig ... end

lexicon "app.bsky.unspecced.defs"

module App_Bsky_Notification_UpdateSeen : sig ... end

lexicon "app.bsky.notification.updateSeen"

module App_Bsky_Notification_RegisterPush : sig ... end

lexicon "app.bsky.notification.registerPush"

module App_Bsky_Notification_PutPreferences : sig ... end

lexicon "app.bsky.notification.putPreferences"

module App_Bsky_Notification_ListNotifications : sig ... end

lexicon "app.bsky.notification.listNotifications"

module App_Bsky_Notification_GetUnreadCount : sig ... end

lexicon "app.bsky.notification.getUnreadCount"

module App_Bsky_Notification_Defs : sig ... end

lexicon "app.bsky.notification.defs"

module App_Bsky_Labeler_Service : sig ... end

lexicon "app.bsky.labeler.service"

module App_Bsky_Labeler_GetServices : sig ... end

lexicon "app.bsky.labeler.getServices"

module App_Bsky_Graph_Verification : sig ... end

lexicon "app.bsky.graph.verification"

module App_Bsky_Graph_UnmuteThread : sig ... end

lexicon "app.bsky.graph.unmuteThread"

module App_Bsky_Graph_UnmuteActorList : sig ... end

lexicon "app.bsky.graph.unmuteActorList"

module App_Bsky_Graph_UnmuteActor : sig ... end

lexicon "app.bsky.graph.unmuteActor"

module App_Bsky_Graph_Starterpack : sig ... end

lexicon "app.bsky.graph.starterpack"

module App_Bsky_Graph_SearchStarterPacks : sig ... end

lexicon "app.bsky.graph.searchStarterPacks"

module App_Bsky_Graph_MuteThread : sig ... end

lexicon "app.bsky.graph.muteThread"

module App_Bsky_Graph_MuteActorList : sig ... end

lexicon "app.bsky.graph.muteActorList"

module App_Bsky_Graph_MuteActor : sig ... end

lexicon "app.bsky.graph.muteActor"

module App_Bsky_Graph_Listitem : sig ... end

lexicon "app.bsky.graph.listitem"

module App_Bsky_Graph_Listblock : sig ... end

lexicon "app.bsky.graph.listblock"

module App_Bsky_Graph_List : sig ... end

lexicon "app.bsky.graph.list"

module App_Bsky_Graph_GetSuggestedFollowsByActor : sig ... end

lexicon "app.bsky.graph.getSuggestedFollowsByActor"

module App_Bsky_Graph_GetStarterPacks : sig ... end

lexicon "app.bsky.graph.getStarterPacks"

module App_Bsky_Graph_GetStarterPack : sig ... end

lexicon "app.bsky.graph.getStarterPack"

module App_Bsky_Graph_GetRelationships : sig ... end

lexicon "app.bsky.graph.getRelationships"

module App_Bsky_Graph_GetMutes : sig ... end

lexicon "app.bsky.graph.getMutes"

module App_Bsky_Graph_GetLists : sig ... end

lexicon "app.bsky.graph.getLists"

module App_Bsky_Graph_GetListMutes : sig ... end

lexicon "app.bsky.graph.getListMutes"

module App_Bsky_Graph_GetListBlocks : sig ... end

lexicon "app.bsky.graph.getListBlocks"

module App_Bsky_Graph_GetList : sig ... end

lexicon "app.bsky.graph.getList"

module App_Bsky_Graph_GetKnownFollowers : sig ... end

lexicon "app.bsky.graph.getKnownFollowers"

module App_Bsky_Graph_GetFollows : sig ... end

lexicon "app.bsky.graph.getFollows"

module App_Bsky_Graph_GetFollowers : sig ... end

lexicon "app.bsky.graph.getFollowers"

module App_Bsky_Graph_GetBlocks : sig ... end

lexicon "app.bsky.graph.getBlocks"

module App_Bsky_Graph_GetActorStarterPacks : sig ... end

lexicon "app.bsky.graph.getActorStarterPacks"

module App_Bsky_Graph_Follow : sig ... end

lexicon "app.bsky.graph.follow"

module App_Bsky_Graph_Block : sig ... end

lexicon "app.bsky.graph.block"

module App_Bsky_Feed_SendInteractions : sig ... end

lexicon "app.bsky.feed.sendInteractions"

module App_Bsky_Feed_SearchPosts : sig ... end

lexicon "app.bsky.feed.searchPosts"

module App_Bsky_Feed_Repost : sig ... end

lexicon "app.bsky.feed.repost"

module App_Bsky_Feed_Post : sig ... end

lexicon "app.bsky.feed.post"

module App_Bsky_Feed_Like : sig ... end

lexicon "app.bsky.feed.like"

module App_Bsky_Feed_GetTimeline : sig ... end

lexicon "app.bsky.feed.getTimeline"

module App_Bsky_Feed_GetSuggestedFeeds : sig ... end

lexicon "app.bsky.feed.getSuggestedFeeds"

module App_Bsky_Feed_GetRepostedBy : sig ... end

lexicon "app.bsky.feed.getRepostedBy"

module App_Bsky_Feed_GetQuotes : sig ... end

lexicon "app.bsky.feed.getQuotes"

module App_Bsky_Feed_GetPosts : sig ... end

lexicon "app.bsky.feed.getPosts"

module App_Bsky_Feed_GetPostThread : sig ... end

lexicon "app.bsky.feed.getPostThread"

module App_Bsky_Feed_GetListFeed : sig ... end

lexicon "app.bsky.feed.getListFeed"

module App_Bsky_Feed_GetLikes : sig ... end

lexicon "app.bsky.feed.getLikes"

module App_Bsky_Feed_GetFeedSkeleton : sig ... end

lexicon "app.bsky.feed.getFeedSkeleton"

module App_Bsky_Feed_GetFeedGenerators : sig ... end

lexicon "app.bsky.feed.getFeedGenerators"

module App_Bsky_Feed_GetFeedGenerator : sig ... end

lexicon "app.bsky.feed.getFeedGenerator"

module App_Bsky_Feed_GetFeed : sig ... end

lexicon "app.bsky.feed.getFeed"

module App_Bsky_Feed_GetAuthorFeed : sig ... end

lexicon "app.bsky.feed.getAuthorFeed"

module App_Bsky_Feed_GetActorLikes : sig ... end

lexicon "app.bsky.feed.getActorLikes"

module App_Bsky_Feed_GetActorFeeds : sig ... end

lexicon "app.bsky.feed.getActorFeeds"

module App_Bsky_Feed_Generator : sig ... end

lexicon "app.bsky.feed.generator"

module App_Bsky_Feed_DescribeFeedGenerator : sig ... end

lexicon "app.bsky.feed.describeFeedGenerator"

module App_Bsky_Actor_Status : sig ... end

lexicon "app.bsky.actor.status"

module App_Bsky_Actor_SearchActorsTypeahead : sig ... end

lexicon "app.bsky.actor.searchActorsTypeahead"

module App_Bsky_Actor_SearchActors : sig ... end

lexicon "app.bsky.actor.searchActors"

module App_Bsky_Actor_PutPreferences : sig ... end

lexicon "app.bsky.actor.putPreferences"

module App_Bsky_Actor_Profile : sig ... end

lexicon "app.bsky.actor.profile"

module App_Bsky_Actor_GetSuggestions : sig ... end

lexicon "app.bsky.actor.getSuggestions"

module App_Bsky_Actor_GetProfiles : sig ... end

lexicon "app.bsky.actor.getProfiles"

module App_Bsky_Actor_GetProfile : sig ... end

lexicon "app.bsky.actor.getProfile"

module App_Bsky_Actor_GetPreferences : sig ... end

lexicon "app.bsky.actor.getPreferences"

module App_Bsky_Actor_Defs : sig ... end

lexicon "app.bsky.actor.defs"

module App_Bsky_Graph_Defs : sig ... end

lexicon "app.bsky.graph.defs"

module App_Bsky_Feed_Defs : sig ... end

lexicon "app.bsky.feed.defs"

module App_Bsky_Embed_RecordWithMedia : sig ... end

lexicon "app.bsky.embed.recordWithMedia" A representation of a record embedded in a Bluesky record (eg, a post), alongside other compatible embeds. For example, a quote post and image, or a quote post and external URL card.

module App_Bsky_Embed_Record : sig ... end

lexicon "app.bsky.embed.record" A representation of a record embedded in a Bluesky record (eg, a post). For example, a quote-post, or sharing a feed generator record.

module App_Bsky_Labeler_Defs : sig ... end

lexicon "app.bsky.labeler.defs"

module Com_Atproto_Repo_StrongRef : sig ... end

lexicon "com.atproto.repo.strongRef" A URI with a content-hash fingerprint.

module Com_Atproto_Moderation_Defs : sig ... end

lexicon "com.atproto.moderation.defs"

module App_Bsky_Embed_Images : sig ... end

lexicon "app.bsky.embed.images" A set of images embedded in a Bluesky record (eg, a post).

module App_Bsky_Embed_Video : sig ... end

lexicon "app.bsky.embed.video" A video embedded in a Bluesky record (eg, a post).

module App_Bsky_Embed_Defs : sig ... end

lexicon "app.bsky.embed.defs"

module App_Bsky_Richtext_Facet : sig ... end

lexicon "app.bsky.richtext.facet"

module Com_Atproto_Label_Defs : sig ... end

lexicon "com.atproto.label.defs"

module App_Bsky_Feed_Threadgate : sig ... end

lexicon "app.bsky.feed.threadgate"

module App_Bsky_Feed_Postgate : sig ... end

lexicon "app.bsky.feed.postgate"

module App_Bsky_Embed_External : sig ... end

lexicon "app.bsky.embed.external"