Module Cephalopod_models.App_Bsky_Actor_Defs

lexicon "app.bsky.actor.defs"

def profileViewBasic

type nonrec profileviewbasic = Types.app_bsky_actor_defs_profileviewbasic = {

did : string;
(*
format: "Did"
*)

handle : string;
(*
format: "Handle"
*)

displayName : string option;
(*
maximum length: 640
*)

avatar : string option;
(*
format: "Uri"
*)

associated : Types.app_bsky_actor_defs_profileassociated option;

viewer : Types.app_bsky_actor_defs_viewerstate option;

labels : Types.com_atproto_label_defs_label list option;

createdAt : string option;
(*
format: "Datetime"
*)

verification : Types.app_bsky_actor_defs_verificationstate option;

status : Types.app_bsky_actor_defs_statusview option;

}

val pp_profileviewbasic : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_profileviewbasic -> Ppx_deriving_runtime.unit

val profileviewbasic_of_value : Types.app_bsky_actor_defs_profileviewbasic Cephalopod_dasl.Value.Util.conv

val profileviewbasic_to_value : Types.app_bsky_actor_defs_profileviewbasic -> Cephalopod_dasl.Value.t

def profileView

type nonrec profileview = Types.app_bsky_actor_defs_profileview = {

did : string;
(*
format: "Did"
*)

handle : string;
(*
format: "Handle"
*)

displayName : string option;
(*
maximum length: 640
*)

description : string option;
(*
maximum length: 2560
*)

avatar : string option;
(*
format: "Uri"
*)

associated : Types.app_bsky_actor_defs_profileassociated option;

indexedAt : string option;
(*
format: "Datetime"
*)

createdAt : string option;
(*
format: "Datetime"
*)

viewer : Types.app_bsky_actor_defs_viewerstate option;

labels : Types.com_atproto_label_defs_label list option;

verification : Types.app_bsky_actor_defs_verificationstate option;

status : Types.app_bsky_actor_defs_statusview option;

}

val pp_profileview : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_profileview -> Ppx_deriving_runtime.unit

val profileview_of_value : Types.app_bsky_actor_defs_profileview Cephalopod_dasl.Value.Util.conv

val profileview_to_value : Types.app_bsky_actor_defs_profileview -> Cephalopod_dasl.Value.t

def profileViewDetailed

type nonrec profileviewdetailed = Types.app_bsky_actor_defs_profileviewdetailed = {

did : string;
(*
format: "Did"
*)

handle : string;
(*
format: "Handle"
*)

displayName : string option;
(*
maximum length: 640
*)

description : string option;
(*
maximum length: 2560
*)

avatar : string option;
(*
format: "Uri"
*)

banner : string option;
(*
format: "Uri"
*)

followersCount : int64 option;

followsCount : int64 option;

postsCount : int64 option;

associated : Types.app_bsky_actor_defs_profileassociated option;

joinedViaStarterPack : Types.app_bsky_graph_defs_starterpackviewbasic option;

indexedAt : string option;
(*
format: "Datetime"
*)

createdAt : string option;
(*
format: "Datetime"
*)

viewer : Types.app_bsky_actor_defs_viewerstate option;

labels : Types.com_atproto_label_defs_label list option;

pinnedPost : Types.com_atproto_repo_strongref_main option;

verification : Types.app_bsky_actor_defs_verificationstate option;

status : Types.app_bsky_actor_defs_statusview option;

}

val pp_profileviewdetailed : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_profileviewdetailed -> Ppx_deriving_runtime.unit

val profileviewdetailed_of_value : Types.app_bsky_actor_defs_profileviewdetailed Cephalopod_dasl.Value.Util.conv

val profileviewdetailed_to_value : Types.app_bsky_actor_defs_profileviewdetailed -> Cephalopod_dasl.Value.t

def profileAssociated

type nonrec profileassociated = Types.app_bsky_actor_defs_profileassociated = {

lists : int64 option;

feedgens : int64 option;

starterPacks : int64 option;

labeler : bool option;

chat : Types.app_bsky_actor_defs_profileassociatedchat option;

}

val pp_profileassociated : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_profileassociated -> Ppx_deriving_runtime.unit

val profileassociated_of_value : Types.app_bsky_actor_defs_profileassociated Cephalopod_dasl.Value.Util.conv

val profileassociated_to_value : Types.app_bsky_actor_defs_profileassociated -> Cephalopod_dasl.Value.t

def profileAssociatedChat

type nonrec profileassociatedchat = Types.app_bsky_actor_defs_profileassociatedchat = {

allowIncoming : string;
(*
known values: "all"; "none"; "following"
*)

}

val pp_profileassociatedchat : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_profileassociatedchat -> Ppx_deriving_runtime.unit

val profileassociatedchat_of_value : Types.app_bsky_actor_defs_profileassociatedchat Cephalopod_dasl.Value.Util.conv

val profileassociatedchat_to_value : Types.app_bsky_actor_defs_profileassociatedchat -> Cephalopod_dasl.Value.t

def viewerState

type nonrec viewerstate = Types.app_bsky_actor_defs_viewerstate = {

muted : bool option;

mutedByList : Types.app_bsky_graph_defs_listviewbasic option;

blockedBy : bool option;

blocking : string option;
(*
format: "AtUri"
*)

blockingByList : Types.app_bsky_graph_defs_listviewbasic option;

following : string option;
(*
format: "AtUri"
*)

followedBy : string option;
(*
format: "AtUri"
*)

knownFollowers : Types.app_bsky_actor_defs_knownfollowers option;

}

val pp_viewerstate : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_viewerstate -> Ppx_deriving_runtime.unit

val viewerstate_of_value : Types.app_bsky_actor_defs_viewerstate Cephalopod_dasl.Value.Util.conv

val viewerstate_to_value : Types.app_bsky_actor_defs_viewerstate -> Cephalopod_dasl.Value.t

def knownFollowers

type nonrec knownfollowers = Types.app_bsky_actor_defs_knownfollowers = {

count : int64;

followers : Types.app_bsky_actor_defs_profileviewbasic list;
(*
maximum length: 5 minimum length: 0
*)

}

val pp_knownfollowers : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_knownfollowers -> Ppx_deriving_runtime.unit

val knownfollowers_of_value : Types.app_bsky_actor_defs_knownfollowers Cephalopod_dasl.Value.Util.conv

val knownfollowers_to_value : Types.app_bsky_actor_defs_knownfollowers -> Cephalopod_dasl.Value.t

def verificationState

type nonrec verificationstate = Types.app_bsky_actor_defs_verificationstate = {

verifications : Types.app_bsky_actor_defs_verificationview list;
(*
All verifications issued by trusted verifiers on behalf of this user. Verifications by untrusted verifiers are not included.
*)

verifiedStatus : string;
(*
The user's status as a verified account. known values: "valid"; "invalid"; "none"
*)

trustedVerifierStatus : string;
(*
The user's status as a trusted verifier. known values: "valid"; "invalid"; "none"
*)

}

val pp_verificationstate : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_verificationstate -> Ppx_deriving_runtime.unit

val verificationstate_of_value : Types.app_bsky_actor_defs_verificationstate Cephalopod_dasl.Value.Util.conv

val verificationstate_to_value : Types.app_bsky_actor_defs_verificationstate -> Cephalopod_dasl.Value.t

def verificationView

type nonrec verificationview = Types.app_bsky_actor_defs_verificationview = {

issuer : string;
(*
The user who issued this verification. format: "Did"
*)

uri : string;
(*
The AT-URI of the verification record. format: "AtUri"
*)

isValid : bool;
(*
True if the verification passes validation, otherwise false.
*)

createdAt : string;
(*
Timestamp when the verification was created. format: "Datetime"
*)

}

val pp_verificationview : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_verificationview -> Ppx_deriving_runtime.unit

val verificationview_of_value : Types.app_bsky_actor_defs_verificationview Cephalopod_dasl.Value.Util.conv

val verificationview_to_value : Types.app_bsky_actor_defs_verificationview -> Cephalopod_dasl.Value.t

def preferences

type nonrec preferences = Types.app_bsky_actor_defs_preferences

val pp_preferences : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_preferences -> Ppx_deriving_runtime.unit

val preferences_of_value : Types.app_bsky_actor_defs_preferences Cephalopod_dasl.Value.Util.conv

val preferences_to_value : [ `App_bsky_actor_defs_adultcontentpref of Types.app_bsky_actor_defs_adultcontentpref | `App_bsky_actor_defs_bskyappstatepref of Types.app_bsky_actor_defs_bskyappstatepref | `App_bsky_actor_defs_contentlabelpref of Types.app_bsky_actor_defs_contentlabelpref | `App_bsky_actor_defs_feedviewpref of Types.app_bsky_actor_defs_feedviewpref | `App_bsky_actor_defs_hiddenpostspref of Types.app_bsky_actor_defs_hiddenpostspref | `App_bsky_actor_defs_interestspref of Types.app_bsky_actor_defs_interestspref | `App_bsky_actor_defs_labelerspref of Types.app_bsky_actor_defs_labelerspref | `App_bsky_actor_defs_mutedwordspref of Types.app_bsky_actor_defs_mutedwordspref | `App_bsky_actor_defs_personaldetailspref of Types.app_bsky_actor_defs_personaldetailspref | `App_bsky_actor_defs_postinteractionsettingspref of Types.app_bsky_actor_defs_postinteractionsettingspref | `App_bsky_actor_defs_savedfeedspref of Types.app_bsky_actor_defs_savedfeedspref | `App_bsky_actor_defs_savedfeedsprefv2 of Types.app_bsky_actor_defs_savedfeedsprefv2 | `App_bsky_actor_defs_threadviewpref of Types.app_bsky_actor_defs_threadviewpref | `App_bsky_actor_defs_verificationprefs of Types.app_bsky_actor_defs_verificationprefs | `Other of Cephalopod_dasl.Value.t ] list -> Cephalopod_dasl.Value.t

def adultContentPref

type nonrec adultcontentpref = Types.app_bsky_actor_defs_adultcontentpref = {

enabled : bool;
(*
default: false
*)

}

val pp_adultcontentpref : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_adultcontentpref -> Ppx_deriving_runtime.unit

val adultcontentpref_of_value : Types.app_bsky_actor_defs_adultcontentpref Cephalopod_dasl.Value.Util.conv

val adultcontentpref_to_value : Types.app_bsky_actor_defs_adultcontentpref -> Cephalopod_dasl.Value.t

def contentLabelPref

type nonrec contentlabelpref = Types.app_bsky_actor_defs_contentlabelpref = {

labelerDid : string option;
(*
Which labeler does this preference apply to? If undefined, applies globally. format: "Did"
*)

label : string;

visibility : string;
(*
known values: "ignore"; "show"; "warn"; "hide"
*)

}

val pp_contentlabelpref : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_contentlabelpref -> Ppx_deriving_runtime.unit

val contentlabelpref_of_value : Types.app_bsky_actor_defs_contentlabelpref Cephalopod_dasl.Value.Util.conv

val contentlabelpref_to_value : Types.app_bsky_actor_defs_contentlabelpref -> Cephalopod_dasl.Value.t

def savedFeed

type nonrec savedfeed = Types.app_bsky_actor_defs_savedfeed = {

id : string;

type_ : string;
(*
known values: "feed"; "list"; "timeline"
*)

value : string;

pinned : bool;

}

val pp_savedfeed : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_savedfeed -> Ppx_deriving_runtime.unit

val savedfeed_of_value : Types.app_bsky_actor_defs_savedfeed Cephalopod_dasl.Value.Util.conv

val savedfeed_to_value : Types.app_bsky_actor_defs_savedfeed -> Cephalopod_dasl.Value.t

def savedFeedsPrefV2

type nonrec savedfeedsprefv2 = Types.app_bsky_actor_defs_savedfeedsprefv2 = {

items : Types.app_bsky_actor_defs_savedfeed list;

}

val pp_savedfeedsprefv2 : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_savedfeedsprefv2 -> Ppx_deriving_runtime.unit

val savedfeedsprefv2_of_value : Types.app_bsky_actor_defs_savedfeedsprefv2 Cephalopod_dasl.Value.Util.conv

val savedfeedsprefv2_to_value : Types.app_bsky_actor_defs_savedfeedsprefv2 -> Cephalopod_dasl.Value.t

def savedFeedsPref

type nonrec savedfeedspref = Types.app_bsky_actor_defs_savedfeedspref = {

pinned : string list;

saved : string list;

timelineIndex : int64 option;

}

val pp_savedfeedspref : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_savedfeedspref -> Ppx_deriving_runtime.unit

val savedfeedspref_of_value : Types.app_bsky_actor_defs_savedfeedspref Cephalopod_dasl.Value.Util.conv

val savedfeedspref_to_value : Types.app_bsky_actor_defs_savedfeedspref -> Cephalopod_dasl.Value.t

def personalDetailsPref

type nonrec personaldetailspref = Types.app_bsky_actor_defs_personaldetailspref = {

birthDate : string option;
(*
The birth date of account owner. format: "Datetime"
*)

}

val pp_personaldetailspref : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_personaldetailspref -> Ppx_deriving_runtime.unit

val personaldetailspref_of_value : Types.app_bsky_actor_defs_personaldetailspref Cephalopod_dasl.Value.Util.conv

val personaldetailspref_to_value : Types.app_bsky_actor_defs_personaldetailspref -> Cephalopod_dasl.Value.t

def feedViewPref

type nonrec feedviewpref = Types.app_bsky_actor_defs_feedviewpref = {

feed : string;
(*
The URI of the feed, or an identifier which describes the feed.
*)

hideReplies : bool option;
(*
Hide replies in the feed.
*)

hideRepliesByUnfollowed : bool option;
(*
Hide replies in the feed if they are not by followed users. default: true
*)

hideRepliesByLikeCount : int64 option;
(*
Hide replies in the feed if they do not have this number of likes.
*)

hideReposts : bool option;
(*
Hide reposts in the feed.
*)

hideQuotePosts : bool option;
(*
Hide quote posts in the feed.
*)

}

val pp_feedviewpref : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_feedviewpref -> Ppx_deriving_runtime.unit

val feedviewpref_of_value : Types.app_bsky_actor_defs_feedviewpref Cephalopod_dasl.Value.Util.conv

val feedviewpref_to_value : Types.app_bsky_actor_defs_feedviewpref -> Cephalopod_dasl.Value.t

def threadViewPref

type nonrec threadviewpref = Types.app_bsky_actor_defs_threadviewpref = {

sort : string option;
(*
Sorting mode for threads. known values: "oldest"; "newest"; "most-likes"; "random"; "hotness"
*)

prioritizeFollowedUsers : bool option;
(*
Show followed users at the top of all replies.
*)

}

val pp_threadviewpref : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_threadviewpref -> Ppx_deriving_runtime.unit

val threadviewpref_of_value : Types.app_bsky_actor_defs_threadviewpref Cephalopod_dasl.Value.Util.conv

val threadviewpref_to_value : Types.app_bsky_actor_defs_threadviewpref -> Cephalopod_dasl.Value.t

def interestsPref

type nonrec interestspref = Types.app_bsky_actor_defs_interestspref = {

tags : string list;
(*
A list of tags which describe the account owner's interests gathered during onboarding. maximum length: 100
*)

}

val pp_interestspref : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_interestspref -> Ppx_deriving_runtime.unit

val interestspref_of_value : Types.app_bsky_actor_defs_interestspref Cephalopod_dasl.Value.Util.conv

val interestspref_to_value : Types.app_bsky_actor_defs_interestspref -> Cephalopod_dasl.Value.t

def mutedWordTarget

type nonrec mutedwordtarget = Types.app_bsky_actor_defs_mutedwordtarget

val pp_mutedwordtarget : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_mutedwordtarget -> Ppx_deriving_runtime.unit

val mutedwordtarget_of_value : Types.app_bsky_actor_defs_mutedwordtarget Cephalopod_dasl.Value.Util.conv

val mutedwordtarget_to_value : string -> Cephalopod_dasl.Value.t

def mutedWord

type nonrec mutedword = Types.app_bsky_actor_defs_mutedword = {

id : string option;

value : string;
(*
The muted word itself. maximum length: 10000
*)

targets : Types.app_bsky_actor_defs_mutedwordtarget list;
(*
The intended targets of the muted word.
*)

actorTarget : string option;
(*
Groups of users to apply the muted word to. If undefined, applies to all users. known values: "all"; "exclude-following"
*)

expiresAt : string option;
(*
The date and time at which the muted word will expire and no longer be applied. format: "Datetime"
*)

}

val pp_mutedword : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_mutedword -> Ppx_deriving_runtime.unit

val mutedword_of_value : Types.app_bsky_actor_defs_mutedword Cephalopod_dasl.Value.Util.conv

val mutedword_to_value : Types.app_bsky_actor_defs_mutedword -> Cephalopod_dasl.Value.t

def mutedWordsPref

type nonrec mutedwordspref = Types.app_bsky_actor_defs_mutedwordspref = {

items : Types.app_bsky_actor_defs_mutedword list;
(*
A list of words the account owner has muted.
*)

}

val pp_mutedwordspref : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_mutedwordspref -> Ppx_deriving_runtime.unit

val mutedwordspref_of_value : Types.app_bsky_actor_defs_mutedwordspref Cephalopod_dasl.Value.Util.conv

val mutedwordspref_to_value : Types.app_bsky_actor_defs_mutedwordspref -> Cephalopod_dasl.Value.t

def hiddenPostsPref

type nonrec hiddenpostspref = Types.app_bsky_actor_defs_hiddenpostspref = {

items : string list;
(*
A list of URIs of posts the account owner has hidden.
*)

}

val pp_hiddenpostspref : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_hiddenpostspref -> Ppx_deriving_runtime.unit

val hiddenpostspref_of_value : Types.app_bsky_actor_defs_hiddenpostspref Cephalopod_dasl.Value.Util.conv

val hiddenpostspref_to_value : Types.app_bsky_actor_defs_hiddenpostspref -> Cephalopod_dasl.Value.t

def labelersPref

type nonrec labelerspref = Types.app_bsky_actor_defs_labelerspref = {

labelers : Types.app_bsky_actor_defs_labelerprefitem list;

}

val pp_labelerspref : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_labelerspref -> Ppx_deriving_runtime.unit

val labelerspref_of_value : Types.app_bsky_actor_defs_labelerspref Cephalopod_dasl.Value.Util.conv

val labelerspref_to_value : Types.app_bsky_actor_defs_labelerspref -> Cephalopod_dasl.Value.t

def labelerPrefItem

type nonrec labelerprefitem = Types.app_bsky_actor_defs_labelerprefitem = {

did : string;
(*
format: "Did"
*)

}

val pp_labelerprefitem : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_labelerprefitem -> Ppx_deriving_runtime.unit

val labelerprefitem_of_value : Types.app_bsky_actor_defs_labelerprefitem Cephalopod_dasl.Value.Util.conv

val labelerprefitem_to_value : Types.app_bsky_actor_defs_labelerprefitem -> Cephalopod_dasl.Value.t

def bskyAppStatePref

type nonrec bskyappstatepref = Types.app_bsky_actor_defs_bskyappstatepref = {

activeProgressGuide : Types.app_bsky_actor_defs_bskyappprogressguide option;

queuedNudges : string list option;
(*
An array of tokens which identify nudges (modals, popups, tours, highlight dots) that should be shown to the user. maximum length: 1000
*)

nuxs : Types.app_bsky_actor_defs_nux list option;
(*
Storage for NUXs the user has encountered. maximum length: 100
*)

}

val pp_bskyappstatepref : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_bskyappstatepref -> Ppx_deriving_runtime.unit

val bskyappstatepref_of_value : Types.app_bsky_actor_defs_bskyappstatepref Cephalopod_dasl.Value.Util.conv

val bskyappstatepref_to_value : Types.app_bsky_actor_defs_bskyappstatepref -> Cephalopod_dasl.Value.t

def bskyAppProgressGuide

type nonrec bskyappprogressguide = Types.app_bsky_actor_defs_bskyappprogressguide = {

guide : string;
(*
maximum length: 100
*)

}

val pp_bskyappprogressguide : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_bskyappprogressguide -> Ppx_deriving_runtime.unit

val bskyappprogressguide_of_value : Types.app_bsky_actor_defs_bskyappprogressguide Cephalopod_dasl.Value.Util.conv

val bskyappprogressguide_to_value : Types.app_bsky_actor_defs_bskyappprogressguide -> Cephalopod_dasl.Value.t

def nux

type nonrec nux = Types.app_bsky_actor_defs_nux = {

id : string;
(*
maximum length: 100
*)

completed : bool;
(*
default: false
*)

data : string option;
(*
Arbitrary data for the NUX. The structure is defined by the NUX itself. Limited to 300 characters. maximum length: 3000
*)

expiresAt : string option;
(*
The date and time at which the NUX will expire and should be considered completed. format: "Datetime"
*)

}

val pp_nux : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_nux -> Ppx_deriving_runtime.unit

val nux_of_value : Types.app_bsky_actor_defs_nux Cephalopod_dasl.Value.Util.conv

val nux_to_value : Types.app_bsky_actor_defs_nux -> Cephalopod_dasl.Value.t

def verificationPrefs

type nonrec verificationprefs = Types.app_bsky_actor_defs_verificationprefs = {

hideBadges : bool option;
(*
Hide the blue check badges for verified accounts and trusted verifiers. default: false
*)

}

val pp_verificationprefs : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_verificationprefs -> Ppx_deriving_runtime.unit

val verificationprefs_of_value : Types.app_bsky_actor_defs_verificationprefs Cephalopod_dasl.Value.Util.conv

val verificationprefs_to_value : Types.app_bsky_actor_defs_verificationprefs -> Cephalopod_dasl.Value.t

def postInteractionSettingsPref

type nonrec postinteractionsettingspref = Types.app_bsky_actor_defs_postinteractionsettingspref = {

threadgateAllowRules : [ `App_bsky_feed_threadgate_mentionrule of Types.app_bsky_feed_threadgate_mentionrule | `App_bsky_feed_threadgate_followerrule of Types.app_bsky_feed_threadgate_followerrule | `App_bsky_feed_threadgate_followingrule of Types.app_bsky_feed_threadgate_followingrule | `App_bsky_feed_threadgate_listrule of Types.app_bsky_feed_threadgate_listrule | `Other of Cephalopod_dasl.Value.t ] list option;
(*
Matches threadgate record. List of rules defining who can reply to this users posts. If value is an empty array, no one can reply. If value is undefined, anyone can reply. maximum length: 5
*)

postgateEmbeddingRules : [ `App_bsky_feed_postgate_disablerule of Types.app_bsky_feed_postgate_disablerule | `Other of Cephalopod_dasl.Value.t ] list option;
(*
Matches postgate record. List of rules defining who can embed this users posts. If value is an empty array or is undefined, no particular rules apply and anyone can embed. maximum length: 5
*)

}

val pp_postinteractionsettingspref : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_postinteractionsettingspref -> Ppx_deriving_runtime.unit

val postinteractionsettingspref_of_value : Types.app_bsky_actor_defs_postinteractionsettingspref Cephalopod_dasl.Value.Util.conv

val postinteractionsettingspref_to_value : Types.app_bsky_actor_defs_postinteractionsettingspref -> Cephalopod_dasl.Value.t

def statusView

type nonrec statusview = Types.app_bsky_actor_defs_statusview = {

status : string;
(*
The status for the account. known values: "app.bsky.actor.status#live"
*)

record : Cephalopod_dasl.Value.t;

embed : [ `App_bsky_embed_external_view of Types.app_bsky_embed_external_view | `Other of Cephalopod_dasl.Value.t ] option;
(*
An optional embed associated with the status.
*)

expiresAt : string option;
(*
The date when this status will expire. The application might choose to no longer return the status after expiration. format: "Datetime"
*)

isActive : bool option;
(*
True if the status is not expired, false if it is expired. Only present if expiration was set.
*)

}

val pp_statusview : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_actor_defs_statusview -> Ppx_deriving_runtime.unit

val statusview_of_value : Types.app_bsky_actor_defs_statusview Cephalopod_dasl.Value.Util.conv

val statusview_to_value : Types.app_bsky_actor_defs_statusview -> Cephalopod_dasl.Value.t