Module Cephalopod_models.App_Bsky_Actor_Profile

lexicon "app.bsky.actor.profile"

def main

type main = {

displayName : string option;
(*
maximum length: 640
*)

description : string option;
(*
Free-form profile description text. maximum length: 2560
*)

avatar : Cephalopod_dasl.Blob.t option;
(*
Small image to be displayed next to posts from account. AKA, 'profile picture' accept: (AcceptN "image/png"; "image/jpeg") max size: 1000000
*)

banner : Cephalopod_dasl.Blob.t option;
(*
Larger horizontal image to display behind profile view. accept: (AcceptN "image/png"; "image/jpeg") max size: 1000000
*)

labels : [ `Com_atproto_label_defs_selflabels of Types.com_atproto_label_defs_selflabels | `Other of Cephalopod_dasl.Value.t ] option;
(*
Self-label values, specific to the Bluesky application, on the overall account.
*)

joinedViaStarterPack : Types.com_atproto_repo_strongref_main option;

pinnedPost : Types.com_atproto_repo_strongref_main option;

createdAt : string option;
(*
format: "Datetime"
*)

}

A declaration of a Bluesky account profile.

val pp_main : Ppx_deriving_runtime.Format.formatter -> main -> Ppx_deriving_runtime.unit

val show_main : main -> Ppx_deriving_runtime.string

val make_main : ?displayName:string -> ?description:string -> ?avatar:Cephalopod_dasl.Blob.t -> ?banner:Cephalopod_dasl.Blob.t -> ?labels: [ `Com_atproto_label_defs_selflabels of Types.com_atproto_label_defs_selflabels | `Other of Cephalopod_dasl.Value.t ] -> ?joinedViaStarterPack:Types.com_atproto_repo_strongref_main -> ?pinnedPost:Types.com_atproto_repo_strongref_main -> ?createdAt:string -> unit -> main

val main_of_value : main Cephalopod_dasl.Value.Util.conv

val main_to_value : main -> Cephalopod_dasl.Value.t

val main : main Base.record_def