Module Cephalopod_models.App_Bsky_Embed_Images

lexicon "app.bsky.embed.images" A set of images embedded in a Bluesky record (eg, a post).

def main

type nonrec main = Types.app_bsky_embed_images_main = {

images : Types.app_bsky_embed_images_image list;
(*
maximum length: 4
*)

}

val pp_main : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_embed_images_main -> Ppx_deriving_runtime.unit

val main_of_value : Types.app_bsky_embed_images_main Cephalopod_dasl.Value.Util.conv

val main_to_value : Types.app_bsky_embed_images_main -> Cephalopod_dasl.Value.t

def image

type nonrec image = Types.app_bsky_embed_images_image = {

image : Cephalopod_dasl.Blob.t;
(*
accept: (AcceptN "image/*") max size: 1000000
*)

alt : string;
(*
Alt text description of the image, for accessibility.
*)

aspectRatio : Types.app_bsky_embed_defs_aspectratio option;

}

val pp_image : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_embed_images_image -> Ppx_deriving_runtime.unit

val image_of_value : Types.app_bsky_embed_images_image Cephalopod_dasl.Value.Util.conv

val image_to_value : Types.app_bsky_embed_images_image -> Cephalopod_dasl.Value.t

def view

type nonrec view = Types.app_bsky_embed_images_view = {

images : Types.app_bsky_embed_images_viewimage list;
(*
maximum length: 4
*)

}

val pp_view : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_embed_images_view -> Ppx_deriving_runtime.unit

val view_of_value : Types.app_bsky_embed_images_view Cephalopod_dasl.Value.Util.conv

val view_to_value : Types.app_bsky_embed_images_view -> Cephalopod_dasl.Value.t

def viewImage

type nonrec viewimage = Types.app_bsky_embed_images_viewimage = {

thumb : string;
(*
Fully-qualified URL where a thumbnail of the image can be fetched. For example, CDN location provided by the App View. format: "Uri"
*)

fullsize : string;
(*
Fully-qualified URL where a large version of the image can be fetched. May or may not be the exact original blob. For example, CDN location provided by the App View. format: "Uri"
*)

alt : string;
(*
Alt text description of the image, for accessibility.
*)

aspectRatio : Types.app_bsky_embed_defs_aspectratio option;

}

val pp_viewimage : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_embed_images_viewimage -> Ppx_deriving_runtime.unit

val viewimage_of_value : Types.app_bsky_embed_images_viewimage Cephalopod_dasl.Value.Util.conv

val viewimage_to_value : Types.app_bsky_embed_images_viewimage -> Cephalopod_dasl.Value.t