Module Cephalopod_models.App_Bsky_Embed_RecordWithMedia

lexicon "app.bsky.embed.recordWithMedia" A representation of a record embedded in a Bluesky record (eg, a post), alongside other compatible embeds. For example, a quote post and image, or a quote post and external URL card.

def main

type nonrec main = Types.app_bsky_embed_recordwithmedia_main = {

record : Types.app_bsky_embed_record_main;

media : [ `App_bsky_embed_images_main of Types.app_bsky_embed_images_main | `App_bsky_embed_video_main of Types.app_bsky_embed_video_main | `App_bsky_embed_external_main of Types.app_bsky_embed_external_main | `Other of Cephalopod_dasl.Value.t ];

}

val pp_main : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_embed_recordwithmedia_main -> Ppx_deriving_runtime.unit

val main_of_value : Types.app_bsky_embed_recordwithmedia_main Cephalopod_dasl.Value.Util.conv

val main_to_value : Types.app_bsky_embed_recordwithmedia_main -> Cephalopod_dasl.Value.t

def view

type nonrec view = Types.app_bsky_embed_recordwithmedia_view = {

record : Types.app_bsky_embed_record_view;

media : [ `App_bsky_embed_images_view of Types.app_bsky_embed_images_view | `App_bsky_embed_video_view of Types.app_bsky_embed_video_view | `App_bsky_embed_external_view of Types.app_bsky_embed_external_view | `Other of Cephalopod_dasl.Value.t ];

}

val pp_view : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_embed_recordwithmedia_view -> Ppx_deriving_runtime.unit

val view_of_value : Types.app_bsky_embed_recordwithmedia_view Cephalopod_dasl.Value.Util.conv

val view_to_value : Types.app_bsky_embed_recordwithmedia_view -> Cephalopod_dasl.Value.t