Module Cephalopod_models.App_Bsky_Feed_Defs

lexicon "app.bsky.feed.defs"

def postView

type nonrec postview = Types.app_bsky_feed_defs_postview = {

uri : string;
(*
format: "AtUri"
*)

cid : string;
(*
format: "Cid"
*)

author : Types.app_bsky_actor_defs_profileviewbasic;

record : Cephalopod_dasl.Value.t;

embed : [ `App_bsky_embed_images_view of Types.app_bsky_embed_images_view | `App_bsky_embed_video_view of Types.app_bsky_embed_video_view | `App_bsky_embed_external_view of Types.app_bsky_embed_external_view | `App_bsky_embed_record_view of Types.app_bsky_embed_record_view | `App_bsky_embed_recordwithmedia_view of Types.app_bsky_embed_recordwithmedia_view | `Other of Cephalopod_dasl.Value.t ] option;

replyCount : int64 option;

repostCount : int64 option;

likeCount : int64 option;

quoteCount : int64 option;

indexedAt : string;
(*
format: "Datetime"
*)

viewer : Types.app_bsky_feed_defs_viewerstate option;

labels : Types.com_atproto_label_defs_label list option;

threadgate : Types.app_bsky_feed_defs_threadgateview option;

}

val pp_postview : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_postview -> Ppx_deriving_runtime.unit

val postview_of_value : Types.app_bsky_feed_defs_postview Cephalopod_dasl.Value.Util.conv

val postview_to_value : Types.app_bsky_feed_defs_postview -> Cephalopod_dasl.Value.t

def viewerState

type nonrec viewerstate = Types.app_bsky_feed_defs_viewerstate = {

repost : string option;
(*
format: "AtUri"
*)

like : string option;
(*
format: "AtUri"
*)

threadMuted : bool option;

replyDisabled : bool option;

embeddingDisabled : bool option;

pinned : bool option;

}

val pp_viewerstate : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_viewerstate -> Ppx_deriving_runtime.unit

val viewerstate_of_value : Types.app_bsky_feed_defs_viewerstate Cephalopod_dasl.Value.Util.conv

val viewerstate_to_value : Types.app_bsky_feed_defs_viewerstate -> Cephalopod_dasl.Value.t

def threadContext

type nonrec threadcontext = Types.app_bsky_feed_defs_threadcontext = {

rootAuthorLike : string option;
(*
format: "AtUri"
*)

}

val pp_threadcontext : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_threadcontext -> Ppx_deriving_runtime.unit

val threadcontext_of_value : Types.app_bsky_feed_defs_threadcontext Cephalopod_dasl.Value.Util.conv

val threadcontext_to_value : Types.app_bsky_feed_defs_threadcontext -> Cephalopod_dasl.Value.t

def feedViewPost

type nonrec feedviewpost = Types.app_bsky_feed_defs_feedviewpost = {

post : Types.app_bsky_feed_defs_postview;

reply : Types.app_bsky_feed_defs_replyref option;

reason : [ `App_bsky_feed_defs_reasonrepost of Types.app_bsky_feed_defs_reasonrepost | `App_bsky_feed_defs_reasonpin of Types.app_bsky_feed_defs_reasonpin | `Other of Cephalopod_dasl.Value.t ] option;

feedContext : string option;
(*
Context provided by feed generator that may be passed back alongside interactions. maximum length: 2000
*)

reqId : string option;
(*
Unique identifier per request that may be passed back alongside interactions. maximum length: 100
*)

}

val pp_feedviewpost : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_feedviewpost -> Ppx_deriving_runtime.unit

val feedviewpost_of_value : Types.app_bsky_feed_defs_feedviewpost Cephalopod_dasl.Value.Util.conv

val feedviewpost_to_value : Types.app_bsky_feed_defs_feedviewpost -> Cephalopod_dasl.Value.t

def replyRef

type nonrec replyref = Types.app_bsky_feed_defs_replyref = {

root : [ `App_bsky_feed_defs_postview of Types.app_bsky_feed_defs_postview | `App_bsky_feed_defs_notfoundpost of Types.app_bsky_feed_defs_notfoundpost | `App_bsky_feed_defs_blockedpost of Types.app_bsky_feed_defs_blockedpost | `Other of Cephalopod_dasl.Value.t ];

parent : [ `App_bsky_feed_defs_postview of Types.app_bsky_feed_defs_postview | `App_bsky_feed_defs_notfoundpost of Types.app_bsky_feed_defs_notfoundpost | `App_bsky_feed_defs_blockedpost of Types.app_bsky_feed_defs_blockedpost | `Other of Cephalopod_dasl.Value.t ];

grandparentAuthor : Types.app_bsky_actor_defs_profileviewbasic option;
(*
When parent is a reply to another post, this is the author of that post.
*)

}

val pp_replyref : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_replyref -> Ppx_deriving_runtime.unit

val replyref_of_value : Types.app_bsky_feed_defs_replyref Cephalopod_dasl.Value.Util.conv

val replyref_to_value : Types.app_bsky_feed_defs_replyref -> Cephalopod_dasl.Value.t

def reasonRepost

type nonrec reasonrepost = Types.app_bsky_feed_defs_reasonrepost = {

by : Types.app_bsky_actor_defs_profileviewbasic;

uri : string option;
(*
format: "AtUri"
*)

cid : string option;
(*
format: "Cid"
*)

indexedAt : string;
(*
format: "Datetime"
*)

}

val pp_reasonrepost : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_reasonrepost -> Ppx_deriving_runtime.unit

val reasonrepost_of_value : Types.app_bsky_feed_defs_reasonrepost Cephalopod_dasl.Value.Util.conv

val reasonrepost_to_value : Types.app_bsky_feed_defs_reasonrepost -> Cephalopod_dasl.Value.t

def reasonPin

type nonrec reasonpin = Types.app_bsky_feed_defs_reasonpin

val pp_reasonpin : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_reasonpin -> Ppx_deriving_runtime.unit

val reasonpin_of_value : Types.app_bsky_feed_defs_reasonpin Cephalopod_dasl.Value.Util.conv

val reasonpin_to_value : Types.app_bsky_feed_defs_reasonpin -> Cephalopod_dasl.Value.t

def threadViewPost

type nonrec threadviewpost = Types.app_bsky_feed_defs_threadviewpost = {

post : Types.app_bsky_feed_defs_postview;

parent : [ `App_bsky_feed_defs_threadviewpost of Types.app_bsky_feed_defs_threadviewpost | `App_bsky_feed_defs_notfoundpost of Types.app_bsky_feed_defs_notfoundpost | `App_bsky_feed_defs_blockedpost of Types.app_bsky_feed_defs_blockedpost | `Other of Cephalopod_dasl.Value.t ] option;

replies : [ `App_bsky_feed_defs_threadviewpost of Types.app_bsky_feed_defs_threadviewpost | `App_bsky_feed_defs_notfoundpost of Types.app_bsky_feed_defs_notfoundpost | `App_bsky_feed_defs_blockedpost of Types.app_bsky_feed_defs_blockedpost | `Other of Cephalopod_dasl.Value.t ] list option;

threadContext : Types.app_bsky_feed_defs_threadcontext option;

}

val pp_threadviewpost : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_threadviewpost -> Ppx_deriving_runtime.unit

val threadviewpost_of_value : Types.app_bsky_feed_defs_threadviewpost Cephalopod_dasl.Value.Util.conv

val threadviewpost_to_value : Types.app_bsky_feed_defs_threadviewpost -> Cephalopod_dasl.Value.t

def notFoundPost

type nonrec notfoundpost = Types.app_bsky_feed_defs_notfoundpost = {

uri : string;
(*
format: "AtUri"
*)

notFound : bool;

}

val pp_notfoundpost : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_notfoundpost -> Ppx_deriving_runtime.unit

val notfoundpost_of_value : Types.app_bsky_feed_defs_notfoundpost Cephalopod_dasl.Value.Util.conv

val notfoundpost_to_value : Types.app_bsky_feed_defs_notfoundpost -> Cephalopod_dasl.Value.t

def blockedPost

type nonrec blockedpost = Types.app_bsky_feed_defs_blockedpost = {

uri : string;
(*
format: "AtUri"
*)

blocked : bool;

author : Types.app_bsky_feed_defs_blockedauthor;

}

val pp_blockedpost : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_blockedpost -> Ppx_deriving_runtime.unit

val blockedpost_of_value : Types.app_bsky_feed_defs_blockedpost Cephalopod_dasl.Value.Util.conv

val blockedpost_to_value : Types.app_bsky_feed_defs_blockedpost -> Cephalopod_dasl.Value.t

def blockedAuthor

type nonrec blockedauthor = Types.app_bsky_feed_defs_blockedauthor = {

did : string;
(*
format: "Did"
*)

viewer : Types.app_bsky_actor_defs_viewerstate option;

}

val pp_blockedauthor : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_blockedauthor -> Ppx_deriving_runtime.unit

val blockedauthor_of_value : Types.app_bsky_feed_defs_blockedauthor Cephalopod_dasl.Value.Util.conv

val blockedauthor_to_value : Types.app_bsky_feed_defs_blockedauthor -> Cephalopod_dasl.Value.t

def generatorView

type nonrec generatorview = Types.app_bsky_feed_defs_generatorview = {

uri : string;
(*
format: "AtUri"
*)

cid : string;
(*
format: "Cid"
*)

did : string;
(*
format: "Did"
*)

creator : Types.app_bsky_actor_defs_profileview;

displayName : string;

description : string option;
(*
maximum length: 3000
*)

descriptionFacets : Types.app_bsky_richtext_facet_main list option;

avatar : string option;
(*
format: "Uri"
*)

likeCount : int64 option;
(*
minimum: 0
*)

acceptsInteractions : bool option;

labels : Types.com_atproto_label_defs_label list option;

viewer : Types.app_bsky_feed_defs_generatorviewerstate option;

contentMode : string option;
(*
known values: "app.bsky.feed.defs#contentModeUnspecified"; "app.bsky.feed.defs#contentModeVideo"
*)

indexedAt : string;
(*
format: "Datetime"
*)

}

val pp_generatorview : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_generatorview -> Ppx_deriving_runtime.unit

val generatorview_of_value : Types.app_bsky_feed_defs_generatorview Cephalopod_dasl.Value.Util.conv

val generatorview_to_value : Types.app_bsky_feed_defs_generatorview -> Cephalopod_dasl.Value.t

def generatorViewerState

type nonrec generatorviewerstate = Types.app_bsky_feed_defs_generatorviewerstate = {

like : string option;
(*
format: "AtUri"
*)

}

val pp_generatorviewerstate : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_generatorviewerstate -> Ppx_deriving_runtime.unit

val generatorviewerstate_of_value : Types.app_bsky_feed_defs_generatorviewerstate Cephalopod_dasl.Value.Util.conv

val generatorviewerstate_to_value : Types.app_bsky_feed_defs_generatorviewerstate -> Cephalopod_dasl.Value.t

def skeletonFeedPost

type nonrec skeletonfeedpost = Types.app_bsky_feed_defs_skeletonfeedpost = {

post : string;
(*
format: "AtUri"
*)

reason : [ `App_bsky_feed_defs_skeletonreasonrepost of Types.app_bsky_feed_defs_skeletonreasonrepost | `App_bsky_feed_defs_skeletonreasonpin of Types.app_bsky_feed_defs_skeletonreasonpin | `Other of Cephalopod_dasl.Value.t ] option;

feedContext : string option;
(*
Context that will be passed through to client and may be passed to feed generator back alongside interactions. maximum length: 2000
*)

}

val pp_skeletonfeedpost : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_skeletonfeedpost -> Ppx_deriving_runtime.unit

val skeletonfeedpost_of_value : Types.app_bsky_feed_defs_skeletonfeedpost Cephalopod_dasl.Value.Util.conv

val skeletonfeedpost_to_value : Types.app_bsky_feed_defs_skeletonfeedpost -> Cephalopod_dasl.Value.t

def skeletonReasonRepost

type nonrec skeletonreasonrepost = Types.app_bsky_feed_defs_skeletonreasonrepost = {

repost : string;
(*
format: "AtUri"
*)

}

val pp_skeletonreasonrepost : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_skeletonreasonrepost -> Ppx_deriving_runtime.unit

val skeletonreasonrepost_of_value : Types.app_bsky_feed_defs_skeletonreasonrepost Cephalopod_dasl.Value.Util.conv

val skeletonreasonrepost_to_value : Types.app_bsky_feed_defs_skeletonreasonrepost -> Cephalopod_dasl.Value.t

def skeletonReasonPin

type nonrec skeletonreasonpin = Types.app_bsky_feed_defs_skeletonreasonpin

val pp_skeletonreasonpin : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_skeletonreasonpin -> Ppx_deriving_runtime.unit

val skeletonreasonpin_of_value : Types.app_bsky_feed_defs_skeletonreasonpin Cephalopod_dasl.Value.Util.conv

val skeletonreasonpin_to_value : Types.app_bsky_feed_defs_skeletonreasonpin -> Cephalopod_dasl.Value.t

def threadgateView

type nonrec threadgateview = Types.app_bsky_feed_defs_threadgateview = {

uri : string option;
(*
format: "AtUri"
*)

cid : string option;
(*
format: "Cid"
*)

record : Cephalopod_dasl.Value.t option;

lists : Types.app_bsky_graph_defs_listviewbasic list option;

}

val pp_threadgateview : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_threadgateview -> Ppx_deriving_runtime.unit

val threadgateview_of_value : Types.app_bsky_feed_defs_threadgateview Cephalopod_dasl.Value.Util.conv

val threadgateview_to_value : Types.app_bsky_feed_defs_threadgateview -> Cephalopod_dasl.Value.t

def interaction

type nonrec interaction = Types.app_bsky_feed_defs_interaction = {

item : string option;
(*
format: "AtUri"
*)

event : string option;
(*
known values: "app.bsky.feed.defs#requestLess"; "app.bsky.feed.defs#requestMore"; "app.bsky.feed.defs#clickthroughItem"; "app.bsky.feed.defs#clickthroughAuthor"; "app.bsky.feed.defs#clickthroughReposter"; "app.bsky.feed.defs#clickthroughEmbed"; "app.bsky.feed.defs#interactionSeen"; "app.bsky.feed.defs#interactionLike"; "app.bsky.feed.defs#interactionRepost"; "app.bsky.feed.defs#interactionReply"; "app.bsky.feed.defs#interactionQuote"; "app.bsky.feed.defs#interactionShare"
*)

feedContext : string option;
(*
Context on a feed item that was originally supplied by the feed generator on getFeedSkeleton. maximum length: 2000
*)

reqId : string option;
(*
Unique identifier per request that may be passed back alongside interactions. maximum length: 100
*)

}

val pp_interaction : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_interaction -> Ppx_deriving_runtime.unit

val interaction_of_value : Types.app_bsky_feed_defs_interaction Cephalopod_dasl.Value.Util.conv

val interaction_to_value : Types.app_bsky_feed_defs_interaction -> Cephalopod_dasl.Value.t

def requestLess

type nonrec requestless = Types.app_bsky_feed_defs_requestless

val pp_requestless : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_requestless -> Ppx_deriving_runtime.unit

val requestless_of_value : Types.app_bsky_feed_defs_requestless Cephalopod_dasl.Value.Util.conv

val requestless_to_value : [ `App_bsky_feed_defs_requestless ] -> Cephalopod_dasl.Value.t

def requestMore

type nonrec requestmore = Types.app_bsky_feed_defs_requestmore

val pp_requestmore : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_requestmore -> Ppx_deriving_runtime.unit

val requestmore_of_value : Types.app_bsky_feed_defs_requestmore Cephalopod_dasl.Value.Util.conv

val requestmore_to_value : [ `App_bsky_feed_defs_requestmore ] -> Cephalopod_dasl.Value.t

def clickthroughItem

type nonrec clickthroughitem = Types.app_bsky_feed_defs_clickthroughitem

val pp_clickthroughitem : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_clickthroughitem -> Ppx_deriving_runtime.unit

val clickthroughitem_of_value : Types.app_bsky_feed_defs_clickthroughitem Cephalopod_dasl.Value.Util.conv

val clickthroughitem_to_value : [ `App_bsky_feed_defs_clickthroughitem ] -> Cephalopod_dasl.Value.t

def clickthroughAuthor

type nonrec clickthroughauthor = Types.app_bsky_feed_defs_clickthroughauthor

val pp_clickthroughauthor : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_clickthroughauthor -> Ppx_deriving_runtime.unit

val clickthroughauthor_of_value : Types.app_bsky_feed_defs_clickthroughauthor Cephalopod_dasl.Value.Util.conv

val clickthroughauthor_to_value : [ `App_bsky_feed_defs_clickthroughauthor ] -> Cephalopod_dasl.Value.t

def clickthroughReposter

type nonrec clickthroughreposter = Types.app_bsky_feed_defs_clickthroughreposter

val pp_clickthroughreposter : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_clickthroughreposter -> Ppx_deriving_runtime.unit

val clickthroughreposter_of_value : Types.app_bsky_feed_defs_clickthroughreposter Cephalopod_dasl.Value.Util.conv

val clickthroughreposter_to_value : [ `App_bsky_feed_defs_clickthroughreposter ] -> Cephalopod_dasl.Value.t

def clickthroughEmbed

type nonrec clickthroughembed = Types.app_bsky_feed_defs_clickthroughembed

val pp_clickthroughembed : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_clickthroughembed -> Ppx_deriving_runtime.unit

val clickthroughembed_of_value : Types.app_bsky_feed_defs_clickthroughembed Cephalopod_dasl.Value.Util.conv

val clickthroughembed_to_value : [ `App_bsky_feed_defs_clickthroughembed ] -> Cephalopod_dasl.Value.t

def contentModeUnspecified

type nonrec contentmodeunspecified = Types.app_bsky_feed_defs_contentmodeunspecified

val pp_contentmodeunspecified : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_contentmodeunspecified -> Ppx_deriving_runtime.unit

val contentmodeunspecified_of_value : Types.app_bsky_feed_defs_contentmodeunspecified Cephalopod_dasl.Value.Util.conv

val contentmodeunspecified_to_value : [ `App_bsky_feed_defs_contentmodeunspecified ] -> Cephalopod_dasl.Value.t

def contentModeVideo

type nonrec contentmodevideo = Types.app_bsky_feed_defs_contentmodevideo

val pp_contentmodevideo : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_contentmodevideo -> Ppx_deriving_runtime.unit

val contentmodevideo_of_value : Types.app_bsky_feed_defs_contentmodevideo Cephalopod_dasl.Value.Util.conv

val contentmodevideo_to_value : [ `App_bsky_feed_defs_contentmodevideo ] -> Cephalopod_dasl.Value.t

def interactionSeen

type nonrec interactionseen = Types.app_bsky_feed_defs_interactionseen

val pp_interactionseen : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_interactionseen -> Ppx_deriving_runtime.unit

val interactionseen_of_value : Types.app_bsky_feed_defs_interactionseen Cephalopod_dasl.Value.Util.conv

val interactionseen_to_value : [ `App_bsky_feed_defs_interactionseen ] -> Cephalopod_dasl.Value.t

def interactionLike

type nonrec interactionlike = Types.app_bsky_feed_defs_interactionlike

val pp_interactionlike : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_interactionlike -> Ppx_deriving_runtime.unit

val interactionlike_of_value : Types.app_bsky_feed_defs_interactionlike Cephalopod_dasl.Value.Util.conv

val interactionlike_to_value : [ `App_bsky_feed_defs_interactionlike ] -> Cephalopod_dasl.Value.t

def interactionRepost

type nonrec interactionrepost = Types.app_bsky_feed_defs_interactionrepost

val pp_interactionrepost : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_interactionrepost -> Ppx_deriving_runtime.unit

val interactionrepost_of_value : Types.app_bsky_feed_defs_interactionrepost Cephalopod_dasl.Value.Util.conv

val interactionrepost_to_value : [ `App_bsky_feed_defs_interactionrepost ] -> Cephalopod_dasl.Value.t

def interactionReply

type nonrec interactionreply = Types.app_bsky_feed_defs_interactionreply

val pp_interactionreply : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_interactionreply -> Ppx_deriving_runtime.unit

val interactionreply_of_value : Types.app_bsky_feed_defs_interactionreply Cephalopod_dasl.Value.Util.conv

val interactionreply_to_value : [ `App_bsky_feed_defs_interactionreply ] -> Cephalopod_dasl.Value.t

def interactionQuote

type nonrec interactionquote = Types.app_bsky_feed_defs_interactionquote

val pp_interactionquote : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_interactionquote -> Ppx_deriving_runtime.unit

val interactionquote_of_value : Types.app_bsky_feed_defs_interactionquote Cephalopod_dasl.Value.Util.conv

val interactionquote_to_value : [ `App_bsky_feed_defs_interactionquote ] -> Cephalopod_dasl.Value.t

def interactionShare

type nonrec interactionshare = Types.app_bsky_feed_defs_interactionshare

val pp_interactionshare : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_defs_interactionshare -> Ppx_deriving_runtime.unit

val interactionshare_of_value : Types.app_bsky_feed_defs_interactionshare Cephalopod_dasl.Value.Util.conv

val interactionshare_to_value : [ `App_bsky_feed_defs_interactionshare ] -> Cephalopod_dasl.Value.t