Module Cephalopod_models.App_Bsky_Feed_Generator

lexicon "app.bsky.feed.generator"

def main

type main = {

did : string;
(*
format: "Did"
*)

displayName : string;
(*
maximum length: 240
*)

description : string option;
(*
maximum length: 3000
*)

descriptionFacets : Types.app_bsky_richtext_facet_main list option;

avatar : Cephalopod_dasl.Blob.t option;
(*
accept: (AcceptN "image/png"; "image/jpeg") max size: 1000000
*)

acceptsInteractions : bool option;
(*
Declaration that a feed accepts feedback interactions from a client through app.bsky.feed.sendInteractions
*)

labels : [ `Com_atproto_label_defs_selflabels of Types.com_atproto_label_defs_selflabels | `Other of Cephalopod_dasl.Value.t ] option;
(*
Self-label values
*)

contentMode : string option;
(*
known values: "app.bsky.feed.defs#contentModeUnspecified"; "app.bsky.feed.defs#contentModeVideo"
*)

createdAt : string;
(*
format: "Datetime"
*)

}

Record declaring of the existence of a feed generator, and containing metadata about it. The record can exist in any repository.

val pp_main : Ppx_deriving_runtime.Format.formatter -> main -> Ppx_deriving_runtime.unit

val show_main : main -> Ppx_deriving_runtime.string

val make_main : did:string -> displayName:string -> ?description:string -> ?descriptionFacets:Types.app_bsky_richtext_facet_main list -> ?avatar:Cephalopod_dasl.Blob.t -> ?acceptsInteractions:bool -> ?labels: [ `Com_atproto_label_defs_selflabels of Types.com_atproto_label_defs_selflabels | `Other of Cephalopod_dasl.Value.t ] -> ?contentMode:string -> createdAt:string -> unit -> main

val main_of_value : main Cephalopod_dasl.Value.Util.conv

val main_to_value : main -> Cephalopod_dasl.Value.t

val main : main Base.record_def