Module Cephalopod_models.App_Bsky_Feed_GetFeedGenerator

lexicon "app.bsky.feed.getFeedGenerator"

def main

type main_params = {

feed : string;
(*
AT-URI of the feed generator record. format: "AtUri"
*)

}

val pp_main_params : Ppx_deriving_runtime.Format.formatter -> main_params -> Ppx_deriving_runtime.unit

val show_main_params : main_params -> Ppx_deriving_runtime.string

val make_main_params : feed:string -> main_params

val main_params_of_value : main_params Cephalopod_dasl.Value.Util.conv

val main_params_to_value : main_params -> Cephalopod_dasl.Value.t

type main_output = {

view : Types.app_bsky_feed_defs_generatorview;

isOnline : bool;
(*
Indicates whether the feed generator service has been online recently, or else seems to be inactive.
*)

isValid : bool;
(*
Indicates whether the feed generator service is compatible with the record declaration.
*)

}

val pp_main_output : Ppx_deriving_runtime.Format.formatter -> main_output -> Ppx_deriving_runtime.unit

val show_main_output : main_output -> Ppx_deriving_runtime.string

val main_output_of_value : main_output Cephalopod_dasl.Value.Util.conv

val main_output_to_value : main_output -> Cephalopod_dasl.Value.t

val main : (main_params, main_output, unit) Base.query

Get information about a feed generator. Implemented by AppView.