Module Cephalopod_models.App_Bsky_Feed_Post

lexicon "app.bsky.feed.post"

def main

type main = {

text : string;
(*
The primary post content. May be an empty string, if there are embeds. maximum length: 3000
*)

entities : Types.app_bsky_feed_post_entity list option;
(*
DEPRECATED: replaced by app.bsky.richtext.facet.
*)

facets : Types.app_bsky_richtext_facet_main list option;
(*
Annotations of text (mentions, URLs, hashtags, etc)
*)

reply : Types.app_bsky_feed_post_replyref option;

embed : [ `App_bsky_embed_images_main of Types.app_bsky_embed_images_main | `App_bsky_embed_video_main of Types.app_bsky_embed_video_main | `App_bsky_embed_external_main of Types.app_bsky_embed_external_main | `App_bsky_embed_record_main of Types.app_bsky_embed_record_main | `App_bsky_embed_recordwithmedia_main of Types.app_bsky_embed_recordwithmedia_main | `Other of Cephalopod_dasl.Value.t ] option;

langs : string list option;
(*
Indicates human language of post primary text content. maximum length: 3
*)

labels : [ `Com_atproto_label_defs_selflabels of Types.com_atproto_label_defs_selflabels | `Other of Cephalopod_dasl.Value.t ] option;
(*
Self-label values for this post. Effectively content warnings.
*)

tags : string list option;
(*
Additional hashtags, in addition to any included in post text and facets. maximum length: 8
*)

createdAt : string;
(*
Client-declared timestamp when this post was originally created. format: "Datetime"
*)

}

Record containing a Bluesky post.

val pp_main : Ppx_deriving_runtime.Format.formatter -> main -> Ppx_deriving_runtime.unit

val show_main : main -> Ppx_deriving_runtime.string

val make_main : text:string -> ?entities:Types.app_bsky_feed_post_entity list -> ?facets:Types.app_bsky_richtext_facet_main list -> ?reply:Types.app_bsky_feed_post_replyref -> ?embed: [ `App_bsky_embed_external_main of Types.app_bsky_embed_external_main | `App_bsky_embed_images_main of Types.app_bsky_embed_images_main | `App_bsky_embed_record_main of Types.app_bsky_embed_record_main | `App_bsky_embed_recordwithmedia_main of Types.app_bsky_embed_recordwithmedia_main | `App_bsky_embed_video_main of Types.app_bsky_embed_video_main | `Other of Cephalopod_dasl.Value.t ] -> ?langs:string list -> ?labels: [ `Com_atproto_label_defs_selflabels of Types.com_atproto_label_defs_selflabels | `Other of Cephalopod_dasl.Value.t ] -> ?tags:string list -> createdAt:string -> unit -> main

val main_of_value : main Cephalopod_dasl.Value.Util.conv

val main_to_value : main -> Cephalopod_dasl.Value.t

val main : main Base.record_def

def replyRef

type nonrec replyref = Types.app_bsky_feed_post_replyref = {

root : Types.com_atproto_repo_strongref_main;

parent : Types.com_atproto_repo_strongref_main;

}

val pp_replyref : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_post_replyref -> Ppx_deriving_runtime.unit

val replyref_of_value : Types.app_bsky_feed_post_replyref Cephalopod_dasl.Value.Util.conv

val replyref_to_value : Types.app_bsky_feed_post_replyref -> Cephalopod_dasl.Value.t

def entity

type nonrec entity = Types.app_bsky_feed_post_entity = {

index : Types.app_bsky_feed_post_textslice;

type_ : string;
(*
Expected values are 'mention' and 'link'.
*)

value : string;

}

val pp_entity : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_post_entity -> Ppx_deriving_runtime.unit

val entity_of_value : Types.app_bsky_feed_post_entity Cephalopod_dasl.Value.Util.conv

val entity_to_value : Types.app_bsky_feed_post_entity -> Cephalopod_dasl.Value.t

def textSlice

type nonrec textslice = Types.app_bsky_feed_post_textslice = {

start : int64;
(*
minimum: 0
*)

end_ : int64;
(*
minimum: 0
*)

}

val pp_textslice : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_post_textslice -> Ppx_deriving_runtime.unit

val textslice_of_value : Types.app_bsky_feed_post_textslice Cephalopod_dasl.Value.Util.conv

val textslice_to_value : Types.app_bsky_feed_post_textslice -> Cephalopod_dasl.Value.t