Module Cephalopod_models.App_Bsky_Feed_Postgate

lexicon "app.bsky.feed.postgate"

def main

type main = {

createdAt : string;
(*
format: "Datetime"
*)

post : string;
(*
Reference (AT-URI) to the post record. format: "AtUri"
*)

detachedEmbeddingUris : string list option;
(*
List of AT-URIs embedding this post that the author has detached from. maximum length: 50
*)

embeddingRules : [ `App_bsky_feed_postgate_disablerule of Types.app_bsky_feed_postgate_disablerule | `Other of Cephalopod_dasl.Value.t ] list option;
(*
List of rules defining who can embed this post. If value is an empty array or is undefined, no particular rules apply and anyone can embed. maximum length: 5
*)

}

Record defining interaction rules for a post. The record key (rkey) of the postgate record must match the record key of the post, and that record must be in the same repository.

val pp_main : Ppx_deriving_runtime.Format.formatter -> main -> Ppx_deriving_runtime.unit

val show_main : main -> Ppx_deriving_runtime.string

val make_main : createdAt:string -> post:string -> ?detachedEmbeddingUris:string list -> ?embeddingRules: [ `App_bsky_feed_postgate_disablerule of Types.app_bsky_feed_postgate_disablerule | `Other of Cephalopod_dasl.Value.t ] list -> unit -> main

val main_of_value : main Cephalopod_dasl.Value.Util.conv

val main_to_value : main -> Cephalopod_dasl.Value.t

val main : main Base.record_def

def disableRule

type nonrec disablerule = Types.app_bsky_feed_postgate_disablerule

val pp_disablerule : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_postgate_disablerule -> Ppx_deriving_runtime.unit

val disablerule_of_value : Types.app_bsky_feed_postgate_disablerule Cephalopod_dasl.Value.Util.conv

val disablerule_to_value : Types.app_bsky_feed_postgate_disablerule -> Cephalopod_dasl.Value.t