Module Cephalopod_models.App_Bsky_Feed_SearchPosts

lexicon "app.bsky.feed.searchPosts"

def main

type main_params = {

q : string;
(*
Search query string; syntax, phrase, boolean, and faceting is unspecified, but Lucene query syntax is recommended.
*)

sort : string option;
(*
Specifies the ranking order of results. known values: "top"; "latest"
*)

since : string option;
(*
Filter results for posts after the indicated datetime (inclusive). Expected to use 'sortAt' timestamp, which may not match 'createdAt'. Can be a datetime, or just an ISO date (YYYY-MM-DD).
*)

until : string option;
(*
Filter results for posts before the indicated datetime (not inclusive). Expected to use 'sortAt' timestamp, which may not match 'createdAt'. Can be a datetime, or just an ISO date (YYY-MM-DD).
*)

mentions : string option;
(*
Filter to posts which mention the given account. Handles are resolved to DID before query-time. Only matches rich-text facet mentions. format: "AtIdentifier"
*)

author : string option;
(*
Filter to posts by the given account. Handles are resolved to DID before query-time. format: "AtIdentifier"
*)

lang : string option;
(*
Filter to posts in the given language. Expected to be based on post language field, though server may override language detection. format: "Language"
*)

domain : string option;
(*
Filter to posts with URLs (facet links or embeds) linking to the given domain (hostname). Server may apply hostname normalization.
*)

url : string option;
(*
Filter to posts with links (facet links or embeds) pointing to this URL. Server may apply URL normalization or fuzzy matching. format: "Uri"
*)

tag : string list option;
(*
Filter to posts with the given tag (hashtag), based on rich-text facet or tag field. Do not include the hash (#) prefix. Multiple tags can be specified, with 'AND' matching.
*)

limit : int64 option;
(*
default: 25 maximum: 100 minimum: 1
*)

cursor : string option;
(*
Optional pagination mechanism; may not necessarily allow scrolling through entire result set.
*)

}

val pp_main_params : Ppx_deriving_runtime.Format.formatter -> main_params -> Ppx_deriving_runtime.unit

val show_main_params : main_params -> Ppx_deriving_runtime.string

val make_main_params : q:string -> ?sort:string -> ?since:string -> ?until:string -> ?mentions:string -> ?author:string -> ?lang:string -> ?domain:string -> ?url:string -> ?tag:string list -> ?limit:int64 -> ?cursor:string -> unit -> main_params

val main_params_of_value : main_params Cephalopod_dasl.Value.Util.conv

val main_params_to_value : main_params -> Cephalopod_dasl.Value.t

type main_output = {

cursor : string option;

hitsTotal : int64 option;
(*
Count of search hits. Optional, may be rounded/truncated, and may not be possible to paginate through all hits.
*)

posts : Types.app_bsky_feed_defs_postview list;

}

val pp_main_output : Ppx_deriving_runtime.Format.formatter -> main_output -> Ppx_deriving_runtime.unit

val show_main_output : main_output -> Ppx_deriving_runtime.string

val main_output_of_value : main_output Cephalopod_dasl.Value.Util.conv

val main_output_to_value : main_output -> Cephalopod_dasl.Value.t

type main_error = [

| `BadQueryString

]

val pp_main_error : Ppx_deriving_runtime.Format.formatter -> main_error -> Ppx_deriving_runtime.unit

val show_main_error : main_error -> Ppx_deriving_runtime.string

val main : (main_params, main_output, main_error) Base.query

Find posts matching search criteria, returning views of those posts. Note that this API endpoint may require authentication (eg, not public) for some service providers and implementations.