Module Cephalopod_models.App_Bsky_Feed_Threadgate

lexicon "app.bsky.feed.threadgate"

def main

type main = {

post : string;
(*
Reference (AT-URI) to the post record. format: "AtUri"
*)

allow : [ `App_bsky_feed_threadgate_mentionrule of Types.app_bsky_feed_threadgate_mentionrule | `App_bsky_feed_threadgate_followerrule of Types.app_bsky_feed_threadgate_followerrule | `App_bsky_feed_threadgate_followingrule of Types.app_bsky_feed_threadgate_followingrule | `App_bsky_feed_threadgate_listrule of Types.app_bsky_feed_threadgate_listrule | `Other of Cephalopod_dasl.Value.t ] list option;
(*
List of rules defining who can reply to this post. If value is an empty array, no one can reply. If value is undefined, anyone can reply. maximum length: 5
*)

createdAt : string;
(*
format: "Datetime"
*)

hiddenReplies : string list option;
(*
List of hidden reply URIs. maximum length: 50
*)

}

Record defining interaction gating rules for a thread (aka, reply controls). The record key (rkey) of the threadgate record must match the record key of the thread's root post, and that record must be in the same repository.

val pp_main : Ppx_deriving_runtime.Format.formatter -> main -> Ppx_deriving_runtime.unit

val show_main : main -> Ppx_deriving_runtime.string

val make_main : post:string -> ?allow: [ `App_bsky_feed_threadgate_followerrule of Types.app_bsky_feed_threadgate_followerrule | `App_bsky_feed_threadgate_followingrule of Types.app_bsky_feed_threadgate_followingrule | `App_bsky_feed_threadgate_listrule of Types.app_bsky_feed_threadgate_listrule | `App_bsky_feed_threadgate_mentionrule of Types.app_bsky_feed_threadgate_mentionrule | `Other of Cephalopod_dasl.Value.t ] list -> createdAt:string -> ?hiddenReplies:string list -> unit -> main

val main_of_value : main Cephalopod_dasl.Value.Util.conv

val main_to_value : main -> Cephalopod_dasl.Value.t

val main : main Base.record_def

def mentionRule

type nonrec mentionrule = Types.app_bsky_feed_threadgate_mentionrule

val pp_mentionrule : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_threadgate_mentionrule -> Ppx_deriving_runtime.unit

val mentionrule_of_value : Types.app_bsky_feed_threadgate_mentionrule Cephalopod_dasl.Value.Util.conv

val mentionrule_to_value : Types.app_bsky_feed_threadgate_mentionrule -> Cephalopod_dasl.Value.t

def followerRule

type nonrec followerrule = Types.app_bsky_feed_threadgate_followerrule

val pp_followerrule : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_threadgate_followerrule -> Ppx_deriving_runtime.unit

val followerrule_of_value : Types.app_bsky_feed_threadgate_followerrule Cephalopod_dasl.Value.Util.conv

val followerrule_to_value : Types.app_bsky_feed_threadgate_followerrule -> Cephalopod_dasl.Value.t

def followingRule

type nonrec followingrule = Types.app_bsky_feed_threadgate_followingrule

val pp_followingrule : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_threadgate_followingrule -> Ppx_deriving_runtime.unit

val followingrule_of_value : Types.app_bsky_feed_threadgate_followingrule Cephalopod_dasl.Value.Util.conv

val followingrule_to_value : Types.app_bsky_feed_threadgate_followingrule -> Cephalopod_dasl.Value.t

def listRule

type nonrec listrule = Types.app_bsky_feed_threadgate_listrule = {

list : string;
(*
format: "AtUri"
*)

}

val pp_listrule : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_feed_threadgate_listrule -> Ppx_deriving_runtime.unit

val listrule_of_value : Types.app_bsky_feed_threadgate_listrule Cephalopod_dasl.Value.Util.conv

val listrule_to_value : Types.app_bsky_feed_threadgate_listrule -> Cephalopod_dasl.Value.t