Module Cephalopod_models.App_Bsky_Labeler_Defs

lexicon "app.bsky.labeler.defs"

def labelerView

type nonrec labelerview = Types.app_bsky_labeler_defs_labelerview = {

uri : string;
(*
format: "AtUri"
*)

cid : string;
(*
format: "Cid"
*)

creator : Types.app_bsky_actor_defs_profileview;

likeCount : int64 option;
(*
minimum: 0
*)

viewer : Types.app_bsky_labeler_defs_labelerviewerstate option;

indexedAt : string;
(*
format: "Datetime"
*)

labels : Types.com_atproto_label_defs_label list option;

}

val pp_labelerview : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_labeler_defs_labelerview -> Ppx_deriving_runtime.unit

val labelerview_of_value : Types.app_bsky_labeler_defs_labelerview Cephalopod_dasl.Value.Util.conv

val labelerview_to_value : Types.app_bsky_labeler_defs_labelerview -> Cephalopod_dasl.Value.t

def labelerViewDetailed

type nonrec labelerviewdetailed = Types.app_bsky_labeler_defs_labelerviewdetailed = {

uri : string;
(*
format: "AtUri"
*)

cid : string;
(*
format: "Cid"
*)

creator : Types.app_bsky_actor_defs_profileview;

policies : Types.app_bsky_labeler_defs_labelerpolicies;

likeCount : int64 option;
(*
minimum: 0
*)

viewer : Types.app_bsky_labeler_defs_labelerviewerstate option;

indexedAt : string;
(*
format: "Datetime"
*)

labels : Types.com_atproto_label_defs_label list option;

reasonTypes : Types.com_atproto_moderation_defs_reasontype list option;
(*
The set of report reason 'codes' which are in-scope for this service to review and action. These usually align to policy categories. If not defined (distinct from empty array), all reason types are allowed.
*)

subjectTypes : Types.com_atproto_moderation_defs_subjecttype list option;
(*
The set of subject types (account, record, etc) this service accepts reports on.
*)

subjectCollections : string list option;
(*
Set of record types (collection NSIDs) which can be reported to this service. If not defined (distinct from empty array), default is any record type.
*)

}

val pp_labelerviewdetailed : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_labeler_defs_labelerviewdetailed -> Ppx_deriving_runtime.unit

val labelerviewdetailed_of_value : Types.app_bsky_labeler_defs_labelerviewdetailed Cephalopod_dasl.Value.Util.conv

val labelerviewdetailed_to_value : Types.app_bsky_labeler_defs_labelerviewdetailed -> Cephalopod_dasl.Value.t

def labelerViewerState

type nonrec labelerviewerstate = Types.app_bsky_labeler_defs_labelerviewerstate = {

like : string option;
(*
format: "AtUri"
*)

}

val pp_labelerviewerstate : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_labeler_defs_labelerviewerstate -> Ppx_deriving_runtime.unit

val labelerviewerstate_of_value : Types.app_bsky_labeler_defs_labelerviewerstate Cephalopod_dasl.Value.Util.conv

val labelerviewerstate_to_value : Types.app_bsky_labeler_defs_labelerviewerstate -> Cephalopod_dasl.Value.t

def labelerPolicies

type nonrec labelerpolicies = Types.app_bsky_labeler_defs_labelerpolicies = {

labelValues : Types.com_atproto_label_defs_labelvalue list;
(*
The label values which this labeler publishes. May include global or custom labels.
*)

labelValueDefinitions : Types.com_atproto_label_defs_labelvaluedefinition list option;
(*
Label values created by this labeler and scoped exclusively to it. Labels defined here will override global label definitions for this labeler.
*)

}

val pp_labelerpolicies : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_labeler_defs_labelerpolicies -> Ppx_deriving_runtime.unit

val labelerpolicies_of_value : Types.app_bsky_labeler_defs_labelerpolicies Cephalopod_dasl.Value.Util.conv

val labelerpolicies_to_value : Types.app_bsky_labeler_defs_labelerpolicies -> Cephalopod_dasl.Value.t