Module Cephalopod_models.App_Bsky_Unspecced_Defs

lexicon "app.bsky.unspecced.defs"

def skeletonSearchPost

type nonrec skeletonsearchpost = Types.app_bsky_unspecced_defs_skeletonsearchpost = {

uri : string;
(*
format: "AtUri"
*)

}

val pp_skeletonsearchpost : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_unspecced_defs_skeletonsearchpost -> Ppx_deriving_runtime.unit

val skeletonsearchpost_of_value : Types.app_bsky_unspecced_defs_skeletonsearchpost Cephalopod_dasl.Value.Util.conv

val skeletonsearchpost_to_value : Types.app_bsky_unspecced_defs_skeletonsearchpost -> Cephalopod_dasl.Value.t

def skeletonSearchActor

type nonrec skeletonsearchactor = Types.app_bsky_unspecced_defs_skeletonsearchactor = {

did : string;
(*
format: "Did"
*)

}

val pp_skeletonsearchactor : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_unspecced_defs_skeletonsearchactor -> Ppx_deriving_runtime.unit

val skeletonsearchactor_of_value : Types.app_bsky_unspecced_defs_skeletonsearchactor Cephalopod_dasl.Value.Util.conv

val skeletonsearchactor_to_value : Types.app_bsky_unspecced_defs_skeletonsearchactor -> Cephalopod_dasl.Value.t

def skeletonSearchStarterPack

type nonrec skeletonsearchstarterpack = Types.app_bsky_unspecced_defs_skeletonsearchstarterpack = {

uri : string;
(*
format: "AtUri"
*)

}

val pp_skeletonsearchstarterpack : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_unspecced_defs_skeletonsearchstarterpack -> Ppx_deriving_runtime.unit

val skeletonsearchstarterpack_of_value : Types.app_bsky_unspecced_defs_skeletonsearchstarterpack Cephalopod_dasl.Value.Util.conv

val skeletonsearchstarterpack_to_value : Types.app_bsky_unspecced_defs_skeletonsearchstarterpack -> Cephalopod_dasl.Value.t

def trendingTopic

type nonrec trendingtopic = Types.app_bsky_unspecced_defs_trendingtopic = {

topic : string;

displayName : string option;

description : string option;

link : string;

}

val pp_trendingtopic : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_unspecced_defs_trendingtopic -> Ppx_deriving_runtime.unit

val trendingtopic_of_value : Types.app_bsky_unspecced_defs_trendingtopic Cephalopod_dasl.Value.Util.conv

val trendingtopic_to_value : Types.app_bsky_unspecced_defs_trendingtopic -> Cephalopod_dasl.Value.t

def skeletonTrend

type nonrec skeletontrend = Types.app_bsky_unspecced_defs_skeletontrend = {

topic : string;

displayName : string;

link : string;

startedAt : string;
(*
format: "Datetime"
*)

postCount : int64;

status : string option;
(*
known values: "hot"
*)

category : string option;

dids : string list;

}

val pp_skeletontrend : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_unspecced_defs_skeletontrend -> Ppx_deriving_runtime.unit

val skeletontrend_of_value : Types.app_bsky_unspecced_defs_skeletontrend Cephalopod_dasl.Value.Util.conv

val skeletontrend_to_value : Types.app_bsky_unspecced_defs_skeletontrend -> Cephalopod_dasl.Value.t

def trendView

type nonrec trendview = Types.app_bsky_unspecced_defs_trendview = {

topic : string;

displayName : string;

link : string;

startedAt : string;
(*
format: "Datetime"
*)

postCount : int64;

status : string option;
(*
known values: "hot"
*)

category : string option;

actors : Types.app_bsky_actor_defs_profileviewbasic list;

}

val pp_trendview : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_unspecced_defs_trendview -> Ppx_deriving_runtime.unit

val trendview_of_value : Types.app_bsky_unspecced_defs_trendview Cephalopod_dasl.Value.Util.conv

val trendview_to_value : Types.app_bsky_unspecced_defs_trendview -> Cephalopod_dasl.Value.t

def threadItemPost

type nonrec threaditempost = Types.app_bsky_unspecced_defs_threaditempost = {

post : Types.app_bsky_feed_defs_postview;

moreParents : bool;
(*
This post has more parents that were not present in the response. This is just a boolean, without the number of parents.
*)

moreReplies : int64;
(*
This post has more replies that were not present in the response. This is a numeric value, which is best-effort and might not be accurate.
*)

opThread : bool;
(*
This post is part of a contiguous thread by the OP from the thread root. Many different OP threads can happen in the same thread.
*)

hiddenByThreadgate : bool;
(*
The threadgate created by the author indicates this post as a reply to be hidden for everyone consuming the thread.
*)

mutedByViewer : bool;
(*
This is by an account muted by the viewer requesting it.
*)

}

val pp_threaditempost : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_unspecced_defs_threaditempost -> Ppx_deriving_runtime.unit

val threaditempost_of_value : Types.app_bsky_unspecced_defs_threaditempost Cephalopod_dasl.Value.Util.conv

val threaditempost_to_value : Types.app_bsky_unspecced_defs_threaditempost -> Cephalopod_dasl.Value.t

def threadItemNoUnauthenticated

type nonrec threaditemnounauthenticated = Types.app_bsky_unspecced_defs_threaditemnounauthenticated

val pp_threaditemnounauthenticated : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_unspecced_defs_threaditemnounauthenticated -> Ppx_deriving_runtime.unit

val threaditemnounauthenticated_of_value : Types.app_bsky_unspecced_defs_threaditemnounauthenticated Cephalopod_dasl.Value.Util.conv

val threaditemnounauthenticated_to_value : Types.app_bsky_unspecced_defs_threaditemnounauthenticated -> Cephalopod_dasl.Value.t

def threadItemNotFound

type nonrec threaditemnotfound = Types.app_bsky_unspecced_defs_threaditemnotfound

val pp_threaditemnotfound : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_unspecced_defs_threaditemnotfound -> Ppx_deriving_runtime.unit

val threaditemnotfound_of_value : Types.app_bsky_unspecced_defs_threaditemnotfound Cephalopod_dasl.Value.Util.conv

val threaditemnotfound_to_value : Types.app_bsky_unspecced_defs_threaditemnotfound -> Cephalopod_dasl.Value.t

def threadItemBlocked

type nonrec threaditemblocked = Types.app_bsky_unspecced_defs_threaditemblocked = {

author : Types.app_bsky_feed_defs_blockedauthor;

}

val pp_threaditemblocked : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_unspecced_defs_threaditemblocked -> Ppx_deriving_runtime.unit

val threaditemblocked_of_value : Types.app_bsky_unspecced_defs_threaditemblocked Cephalopod_dasl.Value.Util.conv

val threaditemblocked_to_value : Types.app_bsky_unspecced_defs_threaditemblocked -> Cephalopod_dasl.Value.t