Module Cephalopod_models.App_Bsky_Unspecced_GetPostThreadHiddenV2

lexicon "app.bsky.unspecced.getPostThreadHiddenV2"

def main

type main_params = {

anchor : string;
(*
Reference (AT-URI) to post record. This is the anchor post. format: "AtUri"
*)

prioritizeFollowedUsers : bool option;
(*
Whether to prioritize posts from followed users. It only has effect when the user is authenticated. default: false
*)

}

val pp_main_params : Ppx_deriving_runtime.Format.formatter -> main_params -> Ppx_deriving_runtime.unit

val show_main_params : main_params -> Ppx_deriving_runtime.string

val make_main_params : anchor:string -> ?prioritizeFollowedUsers:bool -> unit -> main_params

val main_params_of_value : main_params Cephalopod_dasl.Value.Util.conv

val main_params_to_value : main_params -> Cephalopod_dasl.Value.t

type main_output = {

thread : Types.app_bsky_unspecced_getpostthreadhiddenv2_threadhiddenitem list;
(*
A flat list of hidden thread items. The depth of each item is indicated by the depth property inside the item.
*)

}

val pp_main_output : Ppx_deriving_runtime.Format.formatter -> main_output -> Ppx_deriving_runtime.unit

val show_main_output : main_output -> Ppx_deriving_runtime.string

val main_output_of_value : main_output Cephalopod_dasl.Value.Util.conv

val main_output_to_value : main_output -> Cephalopod_dasl.Value.t

val main : (main_params, main_output, unit) Base.query

(NOTE: this endpoint is under development and WILL change without notice. Don't use it until it is moved out of `unspecced` or your application WILL break) Get the hidden posts in a thread. It is based in an anchor post at any depth of the tree, and returns hidden replies (recursive replies, with branching to their replies) below the anchor. It does not include ancestors nor the anchor. This should be called after exhausting `app.bsky.unspecced.getPostThreadV2`. Does not require auth, but additional metadata and filtering will be applied for authed requests.

def threadHiddenItem

type nonrec threadhiddenitem = Types.app_bsky_unspecced_getpostthreadhiddenv2_threadhiddenitem = {

uri : string;
(*
format: "AtUri"
*)

depth : int64;
(*
The nesting level of this item in the thread. Depth 0 means the anchor item. Items above have negative depths, items below have positive depths.
*)

value : [ `App_bsky_unspecced_defs_threaditempost of Types.app_bsky_unspecced_defs_threaditempost | `Other of Cephalopod_dasl.Value.t ];

}

val pp_threadhiddenitem : Ppx_deriving_runtime.Format.formatter -> Types.app_bsky_unspecced_getpostthreadhiddenv2_threadhiddenitem -> Ppx_deriving_runtime.unit

val threadhiddenitem_of_value : Types.app_bsky_unspecced_getpostthreadhiddenv2_threadhiddenitem Cephalopod_dasl.Value.Util.conv

val threadhiddenitem_to_value : Types.app_bsky_unspecced_getpostthreadhiddenv2_threadhiddenitem -> Cephalopod_dasl.Value.t