Module Cephalopod_models.App_Bsky_Unspecced_GetSuggestionsSkeleton

lexicon "app.bsky.unspecced.getSuggestionsSkeleton"

def main

type main_params = {

viewer : string option;
(*
DID of the account making the request (not included for public/unauthenticated queries). Used to boost followed accounts in ranking. format: "Did"
*)

limit : int64 option;
(*
default: 50 maximum: 100 minimum: 1
*)

cursor : string option;

relativeToDid : string option;
(*
DID of the account to get suggestions relative to. If not provided, suggestions will be based on the viewer. format: "Did"
*)

}

val pp_main_params : Ppx_deriving_runtime.Format.formatter -> main_params -> Ppx_deriving_runtime.unit

val show_main_params : main_params -> Ppx_deriving_runtime.string

val make_main_params : ?viewer:string -> ?limit:int64 -> ?cursor:string -> ?relativeToDid:string -> unit -> main_params

val main_params_of_value : main_params Cephalopod_dasl.Value.Util.conv

val main_params_to_value : main_params -> Cephalopod_dasl.Value.t

type main_output = {

cursor : string option;

actors : Types.app_bsky_unspecced_defs_skeletonsearchactor list;

relativeToDid : string option;
(*
DID of the account these suggestions are relative to. If this is returned undefined, suggestions are based on the viewer. format: "Did"
*)

recId : int64 option;
(*
Snowflake for this recommendation, use when submitting recommendation events.
*)

}

val pp_main_output : Ppx_deriving_runtime.Format.formatter -> main_output -> Ppx_deriving_runtime.unit

val show_main_output : main_output -> Ppx_deriving_runtime.string

val main_output_of_value : main_output Cephalopod_dasl.Value.Util.conv

val main_output_to_value : main_output -> Cephalopod_dasl.Value.t

val main : (main_params, main_output, unit) Base.query

Get a skeleton of suggested actors. Intended to be called and then hydrated through app.bsky.actor.getSuggestions