Module Cephalopod_models.Com_Atproto_Identity_GetRecommendedDidCredentials

lexicon "com.atproto.identity.getRecommendedDidCredentials"

def main

type main_output = {

rotationKeys : string list option;
(*
Recommended rotation keys for PLC dids. Should be undefined (or ignored) for did:webs.
*)

alsoKnownAs : string list option;

verificationMethods : Cephalopod_dasl.Value.t option;

services : Cephalopod_dasl.Value.t option;

}

val pp_main_output : Ppx_deriving_runtime.Format.formatter -> main_output -> Ppx_deriving_runtime.unit

val show_main_output : main_output -> Ppx_deriving_runtime.string

val main_output_of_value : main_output Cephalopod_dasl.Value.Util.conv

val main_output_to_value : main_output -> Cephalopod_dasl.Value.t

val main : (unit, main_output, unit) Base.query

Describe the credentials that should be included in the DID doc of an account that is migrating to this service.