Module Cephalopod_models.Com_Atproto_Label_Defs

lexicon "com.atproto.label.defs"

def label

type nonrec label = Types.com_atproto_label_defs_label = {

ver : int64 option;
(*
The AT Protocol version of the label object.
*)

src : string;
(*
DID of the actor who created this label. format: "Did"
*)

uri : string;
(*
AT URI of the record, repository (account), or other resource that this label applies to. format: "Uri"
*)

cid : string option;
(*
Optionally, CID specifying the specific version of 'uri' resource this label applies to. format: "Cid"
*)

val_ : string;
(*
The short string name of the value or type of this label. maximum length: 128
*)

neg : bool option;
(*
If true, this is a negation label, overwriting a previous label.
*)

cts : string;
(*
Timestamp when this label was created. format: "Datetime"
*)

exp : string option;
(*
Timestamp at which this label expires (no longer applies). format: "Datetime"
*)

sig_ : bytes option;
(*
Signature of dag-cbor encoded label.
*)

}

val pp_label : Ppx_deriving_runtime.Format.formatter -> Types.com_atproto_label_defs_label -> Ppx_deriving_runtime.unit

val label_of_value : Types.com_atproto_label_defs_label Cephalopod_dasl.Value.Util.conv

val label_to_value : Types.com_atproto_label_defs_label -> Cephalopod_dasl.Value.t

def selfLabels

type nonrec selflabels = Types.com_atproto_label_defs_selflabels = {

values : Types.com_atproto_label_defs_selflabel list;
(*
maximum length: 10
*)

}

val pp_selflabels : Ppx_deriving_runtime.Format.formatter -> Types.com_atproto_label_defs_selflabels -> Ppx_deriving_runtime.unit

val selflabels_of_value : Types.com_atproto_label_defs_selflabels Cephalopod_dasl.Value.Util.conv

val selflabels_to_value : Types.com_atproto_label_defs_selflabels -> Cephalopod_dasl.Value.t

def selfLabel

type nonrec selflabel = Types.com_atproto_label_defs_selflabel = {

val_ : string;
(*
The short string name of the value or type of this label. maximum length: 128
*)

}

val pp_selflabel : Ppx_deriving_runtime.Format.formatter -> Types.com_atproto_label_defs_selflabel -> Ppx_deriving_runtime.unit

val selflabel_of_value : Types.com_atproto_label_defs_selflabel Cephalopod_dasl.Value.Util.conv

val selflabel_to_value : Types.com_atproto_label_defs_selflabel -> Cephalopod_dasl.Value.t

def labelValueDefinition

type nonrec labelvaluedefinition = Types.com_atproto_label_defs_labelvaluedefinition = {

identifier : string;
(*
The value of the label being defined. Must only include lowercase ascii and the '-' character (a-z-+). maximum length: 100
*)

severity : string;
(*
How should a client visually convey this label? 'inform' means neutral and informational; 'alert' means negative and warning; 'none' means show nothing. known values: "inform"; "alert"; "none"
*)

blurs : string;
(*
What should this label hide in the UI, if applied? 'content' hides all of the target; 'media' hides the images/video/audio; 'none' hides nothing. known values: "content"; "media"; "none"
*)

defaultSetting : string option;
(*
The default setting for this label. known values: "ignore"; "warn"; "hide"
*)

adultOnly : bool option;
(*
Does the user need to have adult content enabled in order to configure this label?
*)

locales : Types.com_atproto_label_defs_labelvaluedefinitionstrings list;

}

val pp_labelvaluedefinition : Ppx_deriving_runtime.Format.formatter -> Types.com_atproto_label_defs_labelvaluedefinition -> Ppx_deriving_runtime.unit

val labelvaluedefinition_of_value : Types.com_atproto_label_defs_labelvaluedefinition Cephalopod_dasl.Value.Util.conv

val labelvaluedefinition_to_value : Types.com_atproto_label_defs_labelvaluedefinition -> Cephalopod_dasl.Value.t

def labelValueDefinitionStrings

type nonrec labelvaluedefinitionstrings = Types.com_atproto_label_defs_labelvaluedefinitionstrings = {

lang : string;
(*
The code of the language these strings are written in. format: "Language"
*)

name : string;
(*
A short human-readable name for the label. maximum length: 640
*)

description : string;
(*
A longer description of what the label means and why it might be applied. maximum length: 100000
*)

}

val pp_labelvaluedefinitionstrings : Ppx_deriving_runtime.Format.formatter -> Types.com_atproto_label_defs_labelvaluedefinitionstrings -> Ppx_deriving_runtime.unit

val labelvaluedefinitionstrings_of_value : Types.com_atproto_label_defs_labelvaluedefinitionstrings Cephalopod_dasl.Value.Util.conv

val labelvaluedefinitionstrings_to_value : Types.com_atproto_label_defs_labelvaluedefinitionstrings -> Cephalopod_dasl.Value.t

def labelValue

type nonrec labelvalue = Types.com_atproto_label_defs_labelvalue

val pp_labelvalue : Ppx_deriving_runtime.Format.formatter -> Types.com_atproto_label_defs_labelvalue -> Ppx_deriving_runtime.unit

val labelvalue_of_value : Types.com_atproto_label_defs_labelvalue Cephalopod_dasl.Value.Util.conv

val labelvalue_to_value : string -> Cephalopod_dasl.Value.t