Module Cephalopod_models.Com_Atproto_Label_QueryLabels

lexicon "com.atproto.label.queryLabels"

def main

type main_params = {

uriPatterns : string list;
(*
List of AT URI patterns to match (boolean 'OR'). Each may be a prefix (ending with '*'; will match inclusive of the string leading to '*'), or a full URI.
*)

sources : string list option;
(*
Optional list of label sources (DIDs) to filter on.
*)

limit : int64 option;
(*
default: 50 maximum: 250 minimum: 1
*)

cursor : string option;

}

val pp_main_params : Ppx_deriving_runtime.Format.formatter -> main_params -> Ppx_deriving_runtime.unit

val show_main_params : main_params -> Ppx_deriving_runtime.string

val make_main_params : ?uriPatterns:string list -> ?sources:string list -> ?limit:int64 -> ?cursor:string -> unit -> main_params

val main_params_of_value : main_params Cephalopod_dasl.Value.Util.conv

val main_params_to_value : main_params -> Cephalopod_dasl.Value.t

type main_output = {

cursor : string option;

labels : Types.com_atproto_label_defs_label list;

}

val pp_main_output : Ppx_deriving_runtime.Format.formatter -> main_output -> Ppx_deriving_runtime.unit

val show_main_output : main_output -> Ppx_deriving_runtime.string

val main_output_of_value : main_output Cephalopod_dasl.Value.Util.conv

val main_output_to_value : main_output -> Cephalopod_dasl.Value.t

val main : (main_params, main_output, unit) Base.query

Find labels relevant to the provided AT-URI patterns. Public endpoint for moderation services, though may return different or additional results with auth.