Module Cephalopod_models.Com_Atproto_Lexicon_Schema

lexicon "com.atproto.lexicon.schema"

def main

type main = {

lexicon : int64;
(*
Indicates the 'version' of the Lexicon language. Must be '1' for the current atproto/Lexicon schema system.
*)

}

Representation of Lexicon schemas themselves, when published as atproto records. Note that the schema language is not defined in Lexicon; this meta schema currently only includes a single version field ('lexicon'). See the atproto specifications for description of the other expected top-level fields ('id', 'defs', etc).

val pp_main : Ppx_deriving_runtime.Format.formatter -> main -> Ppx_deriving_runtime.unit

val show_main : main -> Ppx_deriving_runtime.string

val make_main : lexicon:int64 -> main

val main_of_value : main Cephalopod_dasl.Value.Util.conv

val main_to_value : main -> Cephalopod_dasl.Value.t

val main : main Base.record_def