Module Cephalopod_models.Com_Atproto_Moderation_CreateReport

lexicon "com.atproto.moderation.createReport"

def main

type main_input = {

reasonType : Types.com_atproto_moderation_defs_reasontype;
(*
Indicates the broad category of violation the report is for.
*)

reason : string option;
(*
Additional context about the content and violation. maximum length: 20000
*)

subject : [ `Com_atproto_admin_defs_reporef of Types.com_atproto_admin_defs_reporef | `Com_atproto_repo_strongref_main of Types.com_atproto_repo_strongref_main | `Other of Cephalopod_dasl.Value.t ];

}

val pp_main_input : Ppx_deriving_runtime.Format.formatter -> main_input -> Ppx_deriving_runtime.unit

val show_main_input : main_input -> Ppx_deriving_runtime.string

val main_input_of_value : main_input Cephalopod_dasl.Value.Util.conv

val main_input_to_value : main_input -> Cephalopod_dasl.Value.t

type main_output = {

id : int64;

reasonType : Types.com_atproto_moderation_defs_reasontype;

reason : string option;
(*
maximum length: 20000
*)

subject : [ `Com_atproto_admin_defs_reporef of Types.com_atproto_admin_defs_reporef | `Com_atproto_repo_strongref_main of Types.com_atproto_repo_strongref_main | `Other of Cephalopod_dasl.Value.t ];

reportedBy : string;
(*
format: "Did"
*)

createdAt : string;
(*
format: "Datetime"
*)

}

val pp_main_output : Ppx_deriving_runtime.Format.formatter -> main_output -> Ppx_deriving_runtime.unit

val show_main_output : main_output -> Ppx_deriving_runtime.string

val main_output_of_value : main_output Cephalopod_dasl.Value.Util.conv

val main_output_to_value : main_output -> Cephalopod_dasl.Value.t

val main : (unit, main_input, main_output, unit) Base.procedure

Submit a moderation report regarding an atproto account or record. Implemented by moderation services (with PDS proxying), and requires auth.