Module Cephalopod_models.Com_Atproto_Repo_ApplyWrites

lexicon "com.atproto.repo.applyWrites"

def main

type main_input = {

repo : string;
(*
The handle or DID of the repo (aka, current account). format: "AtIdentifier"
*)

validate : bool option;
(*
Can be set to 'false' to skip Lexicon schema validation of record data across all operations, 'true' to require it, or leave unset to validate only for known Lexicons.
*)

writes : [ `Com_atproto_repo_applywrites_create of Types.com_atproto_repo_applywrites_create | `Com_atproto_repo_applywrites_update of Types.com_atproto_repo_applywrites_update | `Com_atproto_repo_applywrites_delete of Types.com_atproto_repo_applywrites_delete ] list;

swapCommit : string option;
(*
If provided, the entire operation will fail if the current repo commit CID does not match this value. Used to prevent conflicting repo mutations. format: "Cid"
*)

}

val pp_main_input : Ppx_deriving_runtime.Format.formatter -> main_input -> Ppx_deriving_runtime.unit

val show_main_input : main_input -> Ppx_deriving_runtime.string

val main_input_of_value : main_input Cephalopod_dasl.Value.Util.conv

val main_input_to_value : main_input -> Cephalopod_dasl.Value.t

type main_output = {

commit : Types.com_atproto_repo_defs_commitmeta option;

results : [ `Com_atproto_repo_applywrites_createresult of Types.com_atproto_repo_applywrites_createresult | `Com_atproto_repo_applywrites_updateresult of Types.com_atproto_repo_applywrites_updateresult | `Com_atproto_repo_applywrites_deleteresult of Types.com_atproto_repo_applywrites_deleteresult ] list option;

}

val pp_main_output : Ppx_deriving_runtime.Format.formatter -> main_output -> Ppx_deriving_runtime.unit

val show_main_output : main_output -> Ppx_deriving_runtime.string

val main_output_of_value : main_output Cephalopod_dasl.Value.Util.conv

val main_output_to_value : main_output -> Cephalopod_dasl.Value.t

type main_error = [

| `InvalidSwap

]

val pp_main_error : Ppx_deriving_runtime.Format.formatter -> main_error -> Ppx_deriving_runtime.unit

val show_main_error : main_error -> Ppx_deriving_runtime.string

val main : (unit, main_input, main_output, main_error) Base.procedure

Apply a batch transaction of repository creates, updates, and deletes. Requires auth, implemented by PDS.

def create

type nonrec create = Types.com_atproto_repo_applywrites_create = {

collection : string;
(*
format: "Nsid"
*)

rkey : string option;
(*
NOTE: maxLength is redundant with record-key format. Keeping it temporarily to ensure backwards compatibility. format: "RecordKey" maximum length: 512
*)

value : Cephalopod_dasl.Value.t;

}

val pp_create : Ppx_deriving_runtime.Format.formatter -> Types.com_atproto_repo_applywrites_create -> Ppx_deriving_runtime.unit

val create_of_value : Types.com_atproto_repo_applywrites_create Cephalopod_dasl.Value.Util.conv

val create_to_value : Types.com_atproto_repo_applywrites_create -> Cephalopod_dasl.Value.t

def update

type nonrec update = Types.com_atproto_repo_applywrites_update = {

collection : string;
(*
format: "Nsid"
*)

rkey : string;
(*
format: "RecordKey"
*)

value : Cephalopod_dasl.Value.t;

}

val pp_update : Ppx_deriving_runtime.Format.formatter -> Types.com_atproto_repo_applywrites_update -> Ppx_deriving_runtime.unit

val update_of_value : Types.com_atproto_repo_applywrites_update Cephalopod_dasl.Value.Util.conv

val update_to_value : Types.com_atproto_repo_applywrites_update -> Cephalopod_dasl.Value.t

def delete

type nonrec delete = Types.com_atproto_repo_applywrites_delete = {

collection : string;
(*
format: "Nsid"
*)

rkey : string;
(*
format: "RecordKey"
*)

}

val pp_delete : Ppx_deriving_runtime.Format.formatter -> Types.com_atproto_repo_applywrites_delete -> Ppx_deriving_runtime.unit

val delete_of_value : Types.com_atproto_repo_applywrites_delete Cephalopod_dasl.Value.Util.conv

val delete_to_value : Types.com_atproto_repo_applywrites_delete -> Cephalopod_dasl.Value.t

def createResult

type nonrec createresult = Types.com_atproto_repo_applywrites_createresult = {

uri : string;
(*
format: "AtUri"
*)

cid : string;
(*
format: "Cid"
*)

validationStatus : string option;
(*
known values: "valid"; "unknown"
*)

}

val pp_createresult : Ppx_deriving_runtime.Format.formatter -> Types.com_atproto_repo_applywrites_createresult -> Ppx_deriving_runtime.unit

val createresult_of_value : Types.com_atproto_repo_applywrites_createresult Cephalopod_dasl.Value.Util.conv

val createresult_to_value : Types.com_atproto_repo_applywrites_createresult -> Cephalopod_dasl.Value.t

def updateResult

type nonrec updateresult = Types.com_atproto_repo_applywrites_updateresult = {

uri : string;
(*
format: "AtUri"
*)

cid : string;
(*
format: "Cid"
*)

validationStatus : string option;
(*
known values: "valid"; "unknown"
*)

}

val pp_updateresult : Ppx_deriving_runtime.Format.formatter -> Types.com_atproto_repo_applywrites_updateresult -> Ppx_deriving_runtime.unit

val updateresult_of_value : Types.com_atproto_repo_applywrites_updateresult Cephalopod_dasl.Value.Util.conv

val updateresult_to_value : Types.com_atproto_repo_applywrites_updateresult -> Cephalopod_dasl.Value.t

def deleteResult

type nonrec deleteresult = Types.com_atproto_repo_applywrites_deleteresult

val pp_deleteresult : Ppx_deriving_runtime.Format.formatter -> Types.com_atproto_repo_applywrites_deleteresult -> Ppx_deriving_runtime.unit

val deleteresult_of_value : Types.com_atproto_repo_applywrites_deleteresult Cephalopod_dasl.Value.Util.conv

val deleteresult_to_value : Types.com_atproto_repo_applywrites_deleteresult -> Cephalopod_dasl.Value.t