Module Cephalopod_models.Com_Atproto_Repo_ListMissingBlobs

lexicon "com.atproto.repo.listMissingBlobs"

def main

type main_params = {

limit : int64 option;
(*
default: 500 maximum: 1000 minimum: 1
*)

cursor : string option;

}

val pp_main_params : Ppx_deriving_runtime.Format.formatter -> main_params -> Ppx_deriving_runtime.unit

val show_main_params : main_params -> Ppx_deriving_runtime.string

val make_main_params : ?limit:int64 -> ?cursor:string -> unit -> main_params

val main_params_of_value : main_params Cephalopod_dasl.Value.Util.conv

val main_params_to_value : main_params -> Cephalopod_dasl.Value.t

type main_output = {

cursor : string option;

blobs : Types.com_atproto_repo_listmissingblobs_recordblob list;

}

val pp_main_output : Ppx_deriving_runtime.Format.formatter -> main_output -> Ppx_deriving_runtime.unit

val show_main_output : main_output -> Ppx_deriving_runtime.string

val main_output_of_value : main_output Cephalopod_dasl.Value.Util.conv

val main_output_to_value : main_output -> Cephalopod_dasl.Value.t

val main : (main_params, main_output, unit) Base.query

Returns a list of missing blobs for the requesting account. Intended to be used in the account migration flow.

def recordBlob

type nonrec recordblob = Types.com_atproto_repo_listmissingblobs_recordblob = {

cid : string;
(*
format: "Cid"
*)

recordUri : string;
(*
format: "AtUri"
*)

}

val pp_recordblob : Ppx_deriving_runtime.Format.formatter -> Types.com_atproto_repo_listmissingblobs_recordblob -> Ppx_deriving_runtime.unit

val recordblob_of_value : Types.com_atproto_repo_listmissingblobs_recordblob Cephalopod_dasl.Value.Util.conv

val recordblob_to_value : Types.com_atproto_repo_listmissingblobs_recordblob -> Cephalopod_dasl.Value.t