Module Cephalopod_models.Com_Atproto_Repo_PutRecord

lexicon "com.atproto.repo.putRecord"

def main

type main_input = {

repo : string;
(*
The handle or DID of the repo (aka, current account). format: "AtIdentifier"
*)

collection : string;
(*
The NSID of the record collection. format: "Nsid"
*)

rkey : string;
(*
The Record Key. format: "RecordKey" maximum length: 512
*)

validate : bool option;
(*
Can be set to 'false' to skip Lexicon schema validation of record data, 'true' to require it, or leave unset to validate only for known Lexicons.
*)

record : Cephalopod_dasl.Value.t;
(*
The record to write.
*)

swapRecord : string option option;
(*
Compare and swap with the previous record by CID. WARNING: nullable and optional field; may cause problems with golang implementation format: "Cid"
*)

swapCommit : string option;
(*
Compare and swap with the previous commit by CID. format: "Cid"
*)

}

val pp_main_input : Ppx_deriving_runtime.Format.formatter -> main_input -> Ppx_deriving_runtime.unit

val show_main_input : main_input -> Ppx_deriving_runtime.string

val main_input_of_value : main_input Cephalopod_dasl.Value.Util.conv

val main_input_to_value : main_input -> Cephalopod_dasl.Value.t

type main_output = {

uri : string;
(*
format: "AtUri"
*)

cid : string;
(*
format: "Cid"
*)

commit : Types.com_atproto_repo_defs_commitmeta option;

validationStatus : string option;
(*
known values: "valid"; "unknown"
*)

}

val pp_main_output : Ppx_deriving_runtime.Format.formatter -> main_output -> Ppx_deriving_runtime.unit

val show_main_output : main_output -> Ppx_deriving_runtime.string

val main_output_of_value : main_output Cephalopod_dasl.Value.Util.conv

val main_output_to_value : main_output -> Cephalopod_dasl.Value.t

type main_error = [

| `InvalidSwap

]

val pp_main_error : Ppx_deriving_runtime.Format.formatter -> main_error -> Ppx_deriving_runtime.unit

val show_main_error : main_error -> Ppx_deriving_runtime.string

val main : (unit, main_input, main_output, main_error) Base.procedure

Write a repository record, creating or updating it as needed. Requires auth, implemented by PDS.