Module Cephalopod_models.Com_Atproto_Server_CheckAccountStatus

lexicon "com.atproto.server.checkAccountStatus"

def main

type main_output = {

activated : bool;

validDid : bool;

repoCommit : string;
(*
format: "Cid"
*)

repoRev : string;

repoBlocks : int64;

indexedRecords : int64;

privateStateValues : int64;

expectedBlobs : int64;

importedBlobs : int64;

}

val pp_main_output : Ppx_deriving_runtime.Format.formatter -> main_output -> Ppx_deriving_runtime.unit

val show_main_output : main_output -> Ppx_deriving_runtime.string

val main_output_of_value : main_output Cephalopod_dasl.Value.Util.conv

val main_output_to_value : main_output -> Cephalopod_dasl.Value.t

val main : (unit, main_output, unit) Base.query

Returns the status of an account, especially as pertaining to import or recovery. Can be called many times over the course of an account migration. Requires auth and can only be called pertaining to oneself.