Module Cephalopod_models.Com_Atproto_Server_CreateAccount

lexicon "com.atproto.server.createAccount"

def main

type main_input = {

email : string option;

handle : string;
(*
Requested handle for the account. format: "Handle"
*)

did : string option;
(*
Pre-existing atproto DID, being imported to a new account. format: "Did"
*)

inviteCode : string option;

verificationCode : string option;

verificationPhone : string option;

password : string option;
(*
Initial account password. May need to meet instance-specific password strength requirements.
*)

recoveryKey : string option;
(*
DID PLC rotation key (aka, recovery key) to be included in PLC creation operation.
*)

plcOp : Cephalopod_dasl.Value.t option;
(*
A signed DID PLC operation to be submitted as part of importing an existing account to this instance. NOTE: this optional field may be updated when full account migration is implemented.
*)

}

val pp_main_input : Ppx_deriving_runtime.Format.formatter -> main_input -> Ppx_deriving_runtime.unit

val show_main_input : main_input -> Ppx_deriving_runtime.string

val main_input_of_value : main_input Cephalopod_dasl.Value.Util.conv

val main_input_to_value : main_input -> Cephalopod_dasl.Value.t

type main_output = {

accessJwt : string;

refreshJwt : string;

handle : string;
(*
format: "Handle"
*)

did : string;
(*
The DID of the new account. format: "Did"
*)

didDoc : Cephalopod_dasl.Value.t option;
(*
Complete DID document.
*)

}

val pp_main_output : Ppx_deriving_runtime.Format.formatter -> main_output -> Ppx_deriving_runtime.unit

val show_main_output : main_output -> Ppx_deriving_runtime.string

val main_output_of_value : main_output Cephalopod_dasl.Value.Util.conv

val main_output_to_value : main_output -> Cephalopod_dasl.Value.t

type main_error = [

| `InvalidHandle

| `InvalidPassword

| `InvalidInviteCode

| `HandleNotAvailable

| `UnsupportedDomain

| `UnresolvableDid

| `IncompatibleDidDoc

]

val pp_main_error : Ppx_deriving_runtime.Format.formatter -> main_error -> Ppx_deriving_runtime.unit

val show_main_error : main_error -> Ppx_deriving_runtime.string

val main : (unit, main_input, main_output, main_error) Base.procedure

Create an account. Implemented by PDS.