Module Cephalopod_models.Com_Atproto_Server_CreateSession

lexicon "com.atproto.server.createSession"

def main

type main_input = {

identifier : string;
(*
Handle or other identifier supported by the server for the authenticating user.
*)

password : string;

authFactorToken : string option;

allowTakendown : bool option;
(*
When true, instead of throwing error for takendown accounts, a valid response with a narrow scoped token will be returned
*)

}

val pp_main_input : Ppx_deriving_runtime.Format.formatter -> main_input -> Ppx_deriving_runtime.unit

val show_main_input : main_input -> Ppx_deriving_runtime.string

val main_input_of_value : main_input Cephalopod_dasl.Value.Util.conv

val main_input_to_value : main_input -> Cephalopod_dasl.Value.t

type main_output = {

accessJwt : string;

refreshJwt : string;

handle : string;
(*
format: "Handle"
*)

did : string;
(*
format: "Did"
*)

didDoc : Cephalopod_dasl.Value.t option;

email : string option;

emailConfirmed : bool option;

emailAuthFactor : bool option;

active : bool option;

status : string option;
(*
If active=false, this optional field indicates a possible reason for why the account is not active. If active=false and no status is supplied, then the host makes no claim for why the repository is no longer being hosted. known values: "takendown"; "suspended"; "deactivated"
*)

}

val pp_main_output : Ppx_deriving_runtime.Format.formatter -> main_output -> Ppx_deriving_runtime.unit

val show_main_output : main_output -> Ppx_deriving_runtime.string

val main_output_of_value : main_output Cephalopod_dasl.Value.Util.conv

val main_output_to_value : main_output -> Cephalopod_dasl.Value.t

type main_error = [

| `AccountTakedown

| `AuthFactorTokenRequired

]

val pp_main_error : Ppx_deriving_runtime.Format.formatter -> main_error -> Ppx_deriving_runtime.unit

val show_main_error : main_error -> Ppx_deriving_runtime.string

val main : (unit, main_input, main_output, main_error) Base.procedure

Create an authentication session.