Module Cephalopod_models.Com_Atproto_Server_DescribeServer

lexicon "com.atproto.server.describeServer"

def main

type main_output = {

inviteCodeRequired : bool option;
(*
If true, an invite code must be supplied to create an account on this instance.
*)

phoneVerificationRequired : bool option;
(*
If true, a phone verification token must be supplied to create an account on this instance.
*)

availableUserDomains : string list;
(*
List of domain suffixes that can be used in account handles.
*)

links : Types.com_atproto_server_describeserver_links option;
(*
URLs of service policy documents.
*)

contact : Types.com_atproto_server_describeserver_contact option;
(*
Contact information
*)

did : string;
(*
format: "Did"
*)

}

val pp_main_output : Ppx_deriving_runtime.Format.formatter -> main_output -> Ppx_deriving_runtime.unit

val show_main_output : main_output -> Ppx_deriving_runtime.string

val main_output_of_value : main_output Cephalopod_dasl.Value.Util.conv

val main_output_to_value : main_output -> Cephalopod_dasl.Value.t

val main : (unit, main_output, unit) Base.query

Describes the server's account creation requirements and capabilities. Implemented by PDS.

def links

type nonrec links = Types.com_atproto_server_describeserver_links = {

privacyPolicy : string option;
(*
format: "Uri"
*)

termsOfService : string option;
(*
format: "Uri"
*)

}

val pp_links : Ppx_deriving_runtime.Format.formatter -> Types.com_atproto_server_describeserver_links -> Ppx_deriving_runtime.unit

val links_of_value : Types.com_atproto_server_describeserver_links Cephalopod_dasl.Value.Util.conv

val links_to_value : Types.com_atproto_server_describeserver_links -> Cephalopod_dasl.Value.t

def contact

type nonrec contact = Types.com_atproto_server_describeserver_contact = {

email : string option;

}

val pp_contact : Ppx_deriving_runtime.Format.formatter -> Types.com_atproto_server_describeserver_contact -> Ppx_deriving_runtime.unit

val contact_of_value : Types.com_atproto_server_describeserver_contact Cephalopod_dasl.Value.Util.conv

val contact_to_value : Types.com_atproto_server_describeserver_contact -> Cephalopod_dasl.Value.t