Module Cephalopod_models.Com_Atproto_Server_GetServiceAuth

lexicon "com.atproto.server.getServiceAuth"

def main

type main_params = {

aud : string;
(*
The DID of the service that the token will be used to authenticate with format: "Did"
*)

exp : int64 option;
(*
The time in Unix Epoch seconds that the JWT expires. Defaults to 60 seconds in the future. The service may enforce certain time bounds on tokens depending on the requested scope.
*)

lxm : string option;
(*
Lexicon (XRPC) method to bind the requested token to format: "Nsid"
*)

}

val pp_main_params : Ppx_deriving_runtime.Format.formatter -> main_params -> Ppx_deriving_runtime.unit

val show_main_params : main_params -> Ppx_deriving_runtime.string

val make_main_params : aud:string -> ?exp:int64 -> ?lxm:string -> unit -> main_params

val main_params_of_value : main_params Cephalopod_dasl.Value.Util.conv

val main_params_to_value : main_params -> Cephalopod_dasl.Value.t

type main_output = {

token : string;

}

val pp_main_output : Ppx_deriving_runtime.Format.formatter -> main_output -> Ppx_deriving_runtime.unit

val show_main_output : main_output -> Ppx_deriving_runtime.string

val main_output_of_value : main_output Cephalopod_dasl.Value.Util.conv

val main_output_to_value : main_output -> Cephalopod_dasl.Value.t

type main_error = [

| `BadExpiration

]

val pp_main_error : Ppx_deriving_runtime.Format.formatter -> main_error -> Ppx_deriving_runtime.unit

val show_main_error : main_error -> Ppx_deriving_runtime.string

val main : (main_params, main_output, main_error) Base.query

Get a signed token on behalf of the requesting DID for the requested service.