Module Cephalopod_models.Com_Atproto_Server_ReserveSigningKey

lexicon "com.atproto.server.reserveSigningKey"

def main

type main_input = {

did : string option;
(*
The DID to reserve a key for. format: "Did"
*)

}

val pp_main_input : Ppx_deriving_runtime.Format.formatter -> main_input -> Ppx_deriving_runtime.unit

val show_main_input : main_input -> Ppx_deriving_runtime.string

val main_input_of_value : main_input Cephalopod_dasl.Value.Util.conv

val main_input_to_value : main_input -> Cephalopod_dasl.Value.t

type main_output = {

signingKey : string;
(*
The public key for the reserved signing key, in did:key serialization.
*)

}

val pp_main_output : Ppx_deriving_runtime.Format.formatter -> main_output -> Ppx_deriving_runtime.unit

val show_main_output : main_output -> Ppx_deriving_runtime.string

val main_output_of_value : main_output Cephalopod_dasl.Value.Util.conv

val main_output_to_value : main_output -> Cephalopod_dasl.Value.t

val main : (unit, main_input, main_output, unit) Base.procedure

Reserve a repo signing key, for use with account creation. Necessary so that a DID PLC update operation can be constructed during an account migraiton. Public and does not require auth; implemented by PDS. NOTE: this endpoint may change when full account migration is implemented.