Module Cephalopod_models.Com_Atproto_Sync_ListRepos

lexicon "com.atproto.sync.listRepos"

def main

type main_params = {

limit : int64 option;
(*
default: 500 maximum: 1000 minimum: 1
*)

cursor : string option;

}

val pp_main_params : Ppx_deriving_runtime.Format.formatter -> main_params -> Ppx_deriving_runtime.unit

val show_main_params : main_params -> Ppx_deriving_runtime.string

val make_main_params : ?limit:int64 -> ?cursor:string -> unit -> main_params

val main_params_of_value : main_params Cephalopod_dasl.Value.Util.conv

val main_params_to_value : main_params -> Cephalopod_dasl.Value.t

type main_output = {

cursor : string option;

repos : Types.com_atproto_sync_listrepos_repo list;

}

val pp_main_output : Ppx_deriving_runtime.Format.formatter -> main_output -> Ppx_deriving_runtime.unit

val show_main_output : main_output -> Ppx_deriving_runtime.string

val main_output_of_value : main_output Cephalopod_dasl.Value.Util.conv

val main_output_to_value : main_output -> Cephalopod_dasl.Value.t

val main : (main_params, main_output, unit) Base.query

Enumerates all the DID, rev, and commit CID for all repos hosted by this service. Does not require auth; implemented by PDS and Relay.

def repo

type nonrec repo = Types.com_atproto_sync_listrepos_repo = {

did : string;
(*
format: "Did"
*)

head : string;
(*
Current repo commit CID format: "Cid"
*)

rev : string;
(*
format: "Tid"
*)

active : bool option;

status : string option;
(*
If active=false, this optional field indicates a possible reason for why the account is not active. If active=false and no status is supplied, then the host makes no claim for why the repository is no longer being hosted. known values: "takendown"; "suspended"; "deleted"; "deactivated"; "desynchronized"; "throttled"
*)

}

val pp_repo : Ppx_deriving_runtime.Format.formatter -> Types.com_atproto_sync_listrepos_repo -> Ppx_deriving_runtime.unit

val repo_of_value : Types.com_atproto_sync_listrepos_repo Cephalopod_dasl.Value.Util.conv

val repo_to_value : Types.com_atproto_sync_listrepos_repo -> Cephalopod_dasl.Value.t