Module Cephalopod_models.Com_Atproto_Sync_SubscribeRepos

lexicon "com.atproto.sync.subscribeRepos"

def main

type main_msg = [

| `Com_atproto_sync_subscriberepos_commit of Types.com_atproto_sync_subscriberepos_commit

| `Com_atproto_sync_subscriberepos_sync of Types.com_atproto_sync_subscriberepos_sync

| `Com_atproto_sync_subscriberepos_identity of Types.com_atproto_sync_subscriberepos_identity

| `Com_atproto_sync_subscriberepos_account of Types.com_atproto_sync_subscriberepos_account

| `Com_atproto_sync_subscriberepos_info of Types.com_atproto_sync_subscriberepos_info

| `Other of Cephalopod_dasl.Value.t
(*
Non closed union
*)

]

val pp_main_msg : Ppx_deriving_runtime.Format.formatter -> main_msg -> Ppx_deriving_runtime.unit

val show_main_msg : main_msg -> Ppx_deriving_runtime.string

val main_msg_nsid_of_fragment : string -> string

Turn a fragment (in header) into a full nsid

val main_msg_of_value : type_tag:string -> main_msg Cephalopod_dasl.Value.Util.conv

val main_msg_to_value : main_msg -> Cephalopod_dasl.Value.t

type main_params = {

cursor : int64 option;
(*
The last known event seq number to backfill from.
*)

}

val pp_main_params : Ppx_deriving_runtime.Format.formatter -> main_params -> Ppx_deriving_runtime.unit

val show_main_params : main_params -> Ppx_deriving_runtime.string

val make_main_params : ?cursor:int64 -> unit -> main_params

val main_params_of_value : main_params Cephalopod_dasl.Value.Util.conv

val main_params_to_value : main_params -> Cephalopod_dasl.Value.t

type main_error = [

| `FutureCursor

| `ConsumerTooSlow

]

val pp_main_error : Ppx_deriving_runtime.Format.formatter -> main_error -> Ppx_deriving_runtime.unit

val show_main_error : main_error -> Ppx_deriving_runtime.string

val main : (main_params, main_msg, main_error) Base.subscription

Repository event stream, aka Firehose endpoint. Outputs repo commits with diff data, and identity update events, for all repositories on the current server. See the atproto specifications for details around stream sequencing, repo versioning, CAR diff format, and more. Public and does not require auth; implemented by PDS and Relay.

def commit

type nonrec commit = Types.com_atproto_sync_subscriberepos_commit = {

seq : int64;
(*
The stream sequence number of this message.
*)

rebase : bool;
(*
DEPRECATED -- unused
*)

tooBig : bool;
(*
DEPRECATED -- replaced by #sync event and data limits. Indicates that this commit contained too many ops, or data size was too large. Consumers will need to make a separate request to get missing data.
*)

repo : string;
(*
The repo this event comes from. Note that all other message types name this field 'did'. format: "Did"
*)

commit : Cephalopod_dasl.Cid.t;
(*
Repo commit object CID.
*)

rev : string;
(*
The rev of the emitted commit. Note that this information is also in the commit object included in blocks, unless this is a tooBig event. format: "Tid"
*)

since : string option;
(*
The rev of the last emitted commit from this repo (if any). format: "Tid"
*)

blocks : bytes;
(*
CAR file containing relevant blocks, as a diff since the previous repo state. The commit must be included as a block, and the commit block CID must be the first entry in the CAR header 'roots' list. maximum length: 2000000
*)

ops : Types.com_atproto_sync_subscriberepos_repoop list;
(*
maximum length: 200
*)

blobs : Cephalopod_dasl.Cid.t list;

prevData : Cephalopod_dasl.Cid.t option;
(*
The root CID of the MST tree for the previous commit from this repo (indicated by the 'since' revision field in this message). Corresponds to the 'data' field in the repo commit object. NOTE: this field is effectively required for the 'inductive' version of firehose.
*)

time : string;
(*
Timestamp of when this message was originally broadcast. format: "Datetime"
*)

}

val pp_commit : Ppx_deriving_runtime.Format.formatter -> Types.com_atproto_sync_subscriberepos_commit -> Ppx_deriving_runtime.unit

val commit_of_value : Types.com_atproto_sync_subscriberepos_commit Cephalopod_dasl.Value.Util.conv

val commit_to_value : Types.com_atproto_sync_subscriberepos_commit -> Cephalopod_dasl.Value.t

def sync

type nonrec sync = Types.com_atproto_sync_subscriberepos_sync = {

seq : int64;
(*
The stream sequence number of this message.
*)

did : string;
(*
The account this repo event corresponds to. Must match that in the commit object. format: "Did"
*)

blocks : bytes;
(*
CAR file containing the commit, as a block. The CAR header must include the commit block CID as the first 'root'. maximum length: 10000
*)

rev : string;
(*
The rev of the commit. This value must match that in the commit object.
*)

time : string;
(*
Timestamp of when this message was originally broadcast. format: "Datetime"
*)

}

val pp_sync : Ppx_deriving_runtime.Format.formatter -> Types.com_atproto_sync_subscriberepos_sync -> Ppx_deriving_runtime.unit

val sync_of_value : Types.com_atproto_sync_subscriberepos_sync Cephalopod_dasl.Value.Util.conv

val sync_to_value : Types.com_atproto_sync_subscriberepos_sync -> Cephalopod_dasl.Value.t

def identity

type nonrec identity = Types.com_atproto_sync_subscriberepos_identity = {

seq : int64;

did : string;
(*
format: "Did"
*)

time : string;
(*
format: "Datetime"
*)

handle : string option;
(*
The current handle for the account, or 'handle.invalid' if validation fails. This field is optional, might have been validated or passed-through from an upstream source. Semantics and behaviors for PDS vs Relay may evolve in the future; see atproto specs for more details. format: "Handle"
*)

}

val pp_identity : Ppx_deriving_runtime.Format.formatter -> Types.com_atproto_sync_subscriberepos_identity -> Ppx_deriving_runtime.unit

val identity_of_value : Types.com_atproto_sync_subscriberepos_identity Cephalopod_dasl.Value.Util.conv

val identity_to_value : Types.com_atproto_sync_subscriberepos_identity -> Cephalopod_dasl.Value.t

def account

type nonrec account = Types.com_atproto_sync_subscriberepos_account = {

seq : int64;

did : string;
(*
format: "Did"
*)

time : string;
(*
format: "Datetime"
*)

active : bool;
(*
Indicates that the account has a repository which can be fetched from the host that emitted this event.
*)

status : string option;
(*
If active=false, this optional field indicates a reason for why the account is not active. known values: "takendown"; "suspended"; "deleted"; "deactivated"; "desynchronized"; "throttled"
*)

}

val pp_account : Ppx_deriving_runtime.Format.formatter -> Types.com_atproto_sync_subscriberepos_account -> Ppx_deriving_runtime.unit

val account_of_value : Types.com_atproto_sync_subscriberepos_account Cephalopod_dasl.Value.Util.conv

val account_to_value : Types.com_atproto_sync_subscriberepos_account -> Cephalopod_dasl.Value.t

def info

type nonrec info = Types.com_atproto_sync_subscriberepos_info = {

name : string;
(*
known values: "OutdatedCursor"
*)

message : string option;

}

val pp_info : Ppx_deriving_runtime.Format.formatter -> Types.com_atproto_sync_subscriberepos_info -> Ppx_deriving_runtime.unit

val info_of_value : Types.com_atproto_sync_subscriberepos_info Cephalopod_dasl.Value.Util.conv

val info_to_value : Types.com_atproto_sync_subscriberepos_info -> Cephalopod_dasl.Value.t

def repoOp

type nonrec repoop = Types.com_atproto_sync_subscriberepos_repoop = {

action : string;
(*
known values: "create"; "update"; "delete"
*)

path : string;

cid : Cephalopod_dasl.Cid.t option;
(*
For creates and updates, the new record CID. For deletions, null.
*)

prev : Cephalopod_dasl.Cid.t option;
(*
For updates and deletes, the previous record CID (required for inductive firehose). For creations, field should not be defined.
*)

}

val pp_repoop : Ppx_deriving_runtime.Format.formatter -> Types.com_atproto_sync_subscriberepos_repoop -> Ppx_deriving_runtime.unit

val repoop_of_value : Types.com_atproto_sync_subscriberepos_repoop Cephalopod_dasl.Value.Util.conv

val repoop_to_value : Types.com_atproto_sync_subscriberepos_repoop -> Cephalopod_dasl.Value.t